Test.Cardano.Ledger.Shelley.Examples.Consensus

Documentation

type KeyPairWits era = [KeyPair 'Witness (Crypto era)]

data ShelleyResultExamples era

Constructors

ShelleyResultExamples

Fields

srePParams :: PParams era

sreProposedPPUpdates :: ProposedPPUpdates era

srePoolDistr :: PoolDistr (Crypto era)

sreNonMyopicRewards :: Map (Either Coin (Credential 'Staking (Crypto era))) (Map (KeyHash 'StakePool (Crypto era)) Coin)

sreShelleyGenesis :: ShelleyGenesis era

data ShelleyLedgerExamples era

Constructors

ShelleyLedgerExamples

Fields

sleBlock :: Block (BHeader (Crypto era)) era

sleHashHeader :: HashHeader (Crypto era)

sleTx :: Tx era

sleApplyTxError :: ApplyTxError era

sleRewardsCredentials :: Set (Either Coin (Credential 'Staking (Crypto era)))

sleResultExamples :: ShelleyResultExamples era

sleNewEpochState :: NewEpochState era

sleChainDepState :: ChainDepState (Crypto era)

sleTranslationContext :: TranslationContext era

type ShelleyBasedEra' era = (ShelleyBasedEra era, ToCBORGroup (TxSeq era), ToCBOR (Witnesses era), Default (State (EraRule "PPUP" era)), PraosCrypto (Crypto era))

defaultShelleyLedgerExamples :: forall era. (ShelleyBasedEra' era, PredicateFailure (EraRule "DELEGS" era) ~ DelegsPredicateFailure era, PParams era ~ PParams era, PParamsDelta era ~ PParams' StrictMaybe era, Default (StashedAVVMAddresses era)) => (TxBody era -> KeyPairWits era -> Witnesses era) -> (Tx era -> Tx era) -> Value era -> TxBody era -> AuxiliaryData era -> TranslationContext era -> ShelleyLedgerExamples era

exampleShelleyLedgerBlock :: forall era. ShelleyBasedEra' era => Tx era -> Block (BHeader (Crypto era)) era

exampleHashHeader :: forall era. ShelleyBasedEra' era => Proxy era -> HashHeader (Crypto era)

mkKeyHash :: forall c discriminator. Crypto c => Int -> KeyHash discriminator c

mkScriptHash :: forall c. Crypto c => Int -> ScriptHash c

exampleTx :: forall era. ShelleyBasedEra' era => (TxBody era -> KeyPairWits era -> Witnesses era) -> TxBody era -> AuxiliaryData era -> Tx era

This is not a valid transaction. We don't care, we are only interested in serialisation, not validation.

exampleProposedPParamsUpdates :: (ShelleyBasedEra' era, PParamsDelta era ~ PParams' StrictMaybe era) => ProposedPPUpdates era

examplePoolDistr :: forall c. PraosCrypto c => PoolDistr c

exampleNonMyopicRewards :: forall c. PraosCrypto c => Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

testShelleyGenesis :: ShelleyGenesis era

These are dummy values.

exampleNewEpochState :: forall era. (ShelleyBasedEra' era, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_d" (PParams era) UnitInterval, HasField "_protocolVersion" (PParams era) ProtVer, HasField "_nOpt" (PParams era) Natural, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, Default (StashedAVVMAddresses era)) => Value era -> PParams era -> PParams era -> NewEpochState era

This is probably not a valid ledger. We don't care, we are only interested in serialisation, not validation.

exampleLedgerChainDepState :: forall c. PraosCrypto c => Word64 -> ChainDepState c

testEpochInfo :: EpochInfo Identity

mkDummyHash :: forall h a. HashAlgorithm h => Proxy h -> Int -> Hash h a

mkDummySafeHash :: forall c a. Crypto c => Proxy c -> Int -> SafeHash c a

type StandardShelley = ShelleyEra StandardCrypto

ledgerExamplesShelley :: ShelleyLedgerExamples StandardShelley

ShelleyLedgerExamples for Shelley era

mkWitnessesPreAlonzo :: ShelleyBasedEra' era => Proxy era -> TxBody era -> KeyPairWits era -> WitnessSet era

exampleCoin :: Coin

exampleTxBodyShelley :: TxBody StandardShelley

exampleMetadataMap :: Map Word64 Metadatum

exampleAuxiliaryDataShelley :: AuxiliaryData StandardShelley

exampleTxIns :: Crypto c => Set (TxIn c)

exampleCerts :: Crypto c => StrictSeq (DCert c)

exampleWithdrawals :: Crypto c => Wdrl c

exampleProposedPPUpdates :: (PParamsDelta era ~ PParams' StrictMaybe era, ShelleyBasedEra' era) => ProposedPPUpdates era

examplePayKey :: Crypto c => KeyPair 'Payment c

exampleStakeKey :: Crypto c => KeyPair 'Staking c

exampleKeys :: forall c r. Crypto c => AllIssuerKeys c r

keyToCredential :: Crypto c => KeyPair r c -> Credential r c

mkDSIGNKeyPair :: forall c kd. DSIGNAlgorithm (DSIGN c) => Word8 -> KeyPair kd c

mkKeyPair' from Test.Cardano.Ledger.Shelley.Utils doesn't work for real crypto: https://github.com/input-output-hk/cardano-ledger/issues/1770

mkVRFKeyPair :: forall c. VRFAlgorithm (VRF c) => Proxy c -> Word8 -> (SignKeyVRF c, VerKeyVRF c)

examplePoolParams :: forall c. Crypto c => PoolParams c