Test.Cardano.Ledger.Shelley.Examples.Federation

Description

The genesis/core nodes for Shelley Ledger Examples.

Documentation

numCoreNodes :: Word64

Number of Core Node

coreNodeSK :: forall crypto. Crypto crypto => Int -> SignKeyDSIGN crypto

coreNodeVK :: forall crypto. Crypto crypto => Int -> VKey 'Genesis crypto

Verification (Public) Keys

Retrieve the verification key for a core node by providing a number in the range [0, ... (numCoreNodes-1)].

coreNodeIssuerKeys :: forall crypto. Crypto crypto => Int -> AllIssuerKeys crypto 'GenesisDelegate

Block Issuer Keys

Retrieve the block issuer keys (cold, VRF, and hot KES keys) for a core node by providing a number in the range [0, ... (numCoreNodes-1)].

coreNodeKeysBySchedule :: forall era. (HasCallStack, Era era) => PParams era -> Word64 -> AllIssuerKeys (Crypto era) 'GenesisDelegate

Keys by Overlay Schedule

Retrieve all the keys associated with a core node for a given slot and protocol parameters. It will return an error if there is not a core node scheduled for the given slot.

genDelegs :: forall crypto. Crypto crypto => Map (KeyHash 'Genesis crypto) (GenDelegPair crypto)

Genesis Delegation Mapping

The map from genesis/core node (verification) key hashes to their delegate's (verification) key hash.