Test.Cardano.Ledger.Shelley.Generator.Core

Documentation

data AllIssuerKeys v (r ∷ KeyRole)

Constructors

AllIssuerKeys

Fields

aikCold ∷ KeyPair r v

aikVrf ∷ VRFKeyPair v

aikHot ∷ NonEmpty (KESPeriod, KESKeyPair v)

aikColdKeyHash ∷ KeyHash r v

data VRFKeyPair c

Constructors

VRFKeyPair

Fields

vrfSignKey ∷ !(SignKeyVRF c)

vrfVerKey ∷ !(VerKeyVRF c)

data KESKeyPair c

Constructors

KESKeyPair

Fields

kesSignKey ∷ !(SignKeyKES c)

kesVerKey ∷ !(VerKeyKES c)

data GenEnv era

Generator environment.

Constructors

GenEnv

Fields

geKeySpace ∷ KeySpace era

geScriptSpapce ∷ ScriptSpace era

geConstants ∷ Constants

data ScriptSpace era

Constructors

ScriptSpace

Fields

ssScripts3 ∷ [TwoPhase3ArgInfo era]

A list of Two Phase 3 Arg Scripts and their associated data we can use.

ssScripts2 ∷ [TwoPhase2ArgInfo era]

A list of Two Phase 2 Arg Scripts and their associated data we can use.

ssHash3 ∷ Map (ScriptHash (EraCrypto era)) (TwoPhase3ArgInfo era)

ssHash2 ∷ Map (ScriptHash (EraCrypto era)) (TwoPhase2ArgInfo era)

data TwoPhase3ArgInfo era

Constructors

TwoPhase3ArgInfo

Fields

getScript3 ∷ Script era

A Plutus Script

getHash3 ∷ ScriptHash (EraCrypto era)

Its ScriptHash

getData3 ∷ Data

A Data that will make it succeed

getRedeemer3 ∷ (Data, Natural, Natural)

A Redeemer that will make it succeed

getSucceeds3 ∷ Bool

data TwoPhase2ArgInfo era

Constructors

TwoPhase2ArgInfo

Fields

getScript2 ∷ Script era

A Plutus Script

getHash2 ∷ ScriptHash (EraCrypto era)

Its ScriptHash

getRedeemer2 ∷ (Data, Natural, Natural)

A Redeemer that will make it succeed

getSucceeds2 ∷ Bool

type ScriptInfo era = (Map (ScriptHash (EraCrypto era)) (TwoPhase3ArgInfo era), Map (ScriptHash (EraCrypto era)) (TwoPhase2ArgInfo era))

data KeySpace era

Collection of all keys which are required to generate a trace.

These are the _only_ keys which should be involved in the trace.

Constructors

KeySpace_

Fields

ksCoreNodes ∷ [(GenesisKeyPair (EraCrypto era), AllIssuerKeys (EraCrypto era) 'GenesisDelegate)]

ksGenesisDelegates ∷ [AllIssuerKeys (EraCrypto era) 'GenesisDelegate]

Bag of keys to be used for future genesis delegates

ksStakePools ∷ [AllIssuerKeys (EraCrypto era) 'StakePool]

Bag of keys to be used for future stake pools

ksKeyPairs ∷ KeyPairs (EraCrypto era)

Bag of keys to be used for future payment/staking addresses

ksMSigScripts ∷ [(Script era, Script era)]

ksIndexedPaymentKeys ∷ Map (KeyHash 'Payment (EraCrypto era)) (KeyPair 'Payment (EraCrypto era))

Index over the payment keys in ksKeyPairs

ksIndexedStakingKeys ∷ Map (KeyHash 'Staking (EraCrypto era)) (KeyPair 'Staking (EraCrypto era))

Index over the staking keys in ksKeyPairs

ksIndexedGenDelegates ∷ Map (KeyHash 'GenesisDelegate (EraCrypto era)) (AllIssuerKeys (EraCrypto era) 'GenesisDelegate)

Index over the cold key hashes in Genesis Delegates

ksIndexedPayScripts ∷ Map (ScriptHash (EraCrypto era)) (Script era, Script era)

Index over the pay script hashes in Script pairs

ksIndexedStakeScripts ∷ Map (ScriptHash (EraCrypto era)) (Script era, Script era)

Index over the stake script hashes in Script pairs

pattern KeySpace ∷ ∀ era. ScriptClass era ⇒ [(GenesisKeyPair (EraCrypto era), AllIssuerKeys (EraCrypto era) 'GenesisDelegate)] → [AllIssuerKeys (EraCrypto era) 'GenesisDelegate] → [AllIssuerKeys (EraCrypto era) 'StakePool] → KeyPairs (EraCrypto era) → [(Script era, Script era)] → KeySpace era

newtype NatNonce

We provide our own nonces to mkBlock, which we then wish to recover as the output of the VRF functions. In general, however, we just derive them from a natural. Since the nonce is a hash, we do not want to recover it to find a preimage. In testing, therefore, we just wrap the raw natural, which we then encode into the fake VRF implementation.

Constructors

NatNonce Natural

findPayKeyPairAddr ∷ ∀ era. Addr (EraCrypto era) → Map (KeyHash 'Payment (EraCrypto era)) (KeyPair 'Payment (EraCrypto era)) → KeyPair 'Payment (EraCrypto era)

Find first matching key pair for address. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayKeyPairCred ∷ ∀ era kr. Credential kr (EraCrypto era) → Map (KeyHash kr (EraCrypto era)) (KeyPair kr (EraCrypto era)) → KeyPair kr (EraCrypto era)

Find first matching key pair for a credential. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayScriptFromCred ∷ ∀ era. Credential 'Witness (EraCrypto era) → Map (ScriptHash (EraCrypto era)) (Script era, Script era) → (Script era, Script era)

Find matching multisig scripts for a credential.

findStakeScriptFromCred ∷ Credential 'Witness (EraCrypto era) → Map (ScriptHash (EraCrypto era)) (Script era, Script era) → (Script era, Script era)

Find first matching script for a credential.

findPayScriptFromAddr ∷ ∀ era. Addr (EraCrypto era) → Map (ScriptHash (EraCrypto era)) (Script era, Script era) → (Script era, Script era)

Find first matching multisig script for an address.

genBool ∷ Gen Bool

genCoinList ∷ Integer → Integer → Int → Gen [Coin]

Generates a list of Coin values of length between lower and upper and with values between minCoin and maxCoin.

genInteger ∷ Integer → Integer → Gen Integer

genNatural ∷ Natural → Natural → Gen Natural

Generator for a natural number between lower and upper

genWord64 ∷ Word64 → Word64 → Gen Word64

Generator for a Word64 between lower and upper

genTxOut ∷ ∀ era. EraTxOut era ⇒ Gen (Value era) → [Addr (EraCrypto era)] → Gen [TxOut era]

Generates a list of coins for the given Addr and produced a TxOut for each Addr

Note: we need to keep the initial utxo coin sizes large enough so that when we simulate sequences of transactions, we have enough funds available to include certificates that require deposits.

genesisCoins ∷ TxId (EraCrypto era) → [TxOut era] → UTxO era

Creates the UTxO for a new ledger with the specified genesis TxId and transaction outputs.

increasingProbabilityAt ∷ Gen a → (a, a) → Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

pickStakeKey ∷ KeyPairs c → Gen (VKey 'Staking c)

Select one random verification staking key from list of pairs of KeyPair.

mkAddr ∷ Crypto c ⇒ (KeyPair 'Payment c, KeyPair 'Staking c) → Addr c

mkCred ∷ ∀ c (kr ∷ KeyRole). Crypto c ⇒ KeyPair kr c → Credential kr c

unitIntervalToNatural ∷ UnitInterval → Natural

Deprecated: This function has been made private in cardano-protocol-tpraos:testlib. Open an issue if you need it

Try to map the unit interval to a natural number. We don't care whether this is surjective. But it should be right inverse to fromNatural - that is, one should be able to recover the UnitInterval value used here.

mkBlock

Arguments

∷ ∀ era (r ∷ KeyRole). (EraSegWits era, Signable (VRF (EraCrypto era)) Seed, Signable (KES (EraCrypto era)) (BHBody (EraCrypto era)))

⇒ HashHeader (EraCrypto era)
Hash of previous block

→ AllIssuerKeys (EraCrypto era) r
All keys in the stake pool

→ [Tx era]
Transactions to record

→ SlotNo
Current slot

→ BlockNo
Block numberchain lengthchain "difficulty"

→ Nonce
EpochNo nonce

→ Word
Period of KES (key evolving signature scheme)

→ Word
KES period of key registration

→ OCert (EraCrypto era)
Operational certificate

→ Block (BHeader (EraCrypto era)) era

mkBlockHeader

Arguments

∷ Mock c

⇒ ProtVer

→ HashHeader c
Hash of previous block

→ AllIssuerKeys c r
All keys in the stake pool

→ SlotNo
Current slot

→ BlockNo
Block numberchain lengthchain "difficulty"

→ Nonce
EpochNo nonce

→ Word
Period of KES (key evolving signature scheme)

→ Word
KES period of key registration

→ OCert c
Operational certificate

→ Word32
Block size

→ Hash c EraIndependentBlockBody
Block body hash

→ BHeader c

Deprecated: In favor of mkBHeader and mkBHBody

mkBlockFakeVRF

Arguments

∷ ∀ era (r ∷ KeyRole). (EraSegWits era, Signable (VRF (EraCrypto era)) (WithResult Seed), Signable (KES (EraCrypto era)) (BHBody (EraCrypto era)))

⇒ HashHeader (EraCrypto era)
Hash of previous block

→ AllIssuerKeys (EraCrypto era) r
All keys in the stake pool

→ [Tx era]
Transactions to record

→ SlotNo
Current slot

→ BlockNo
Block number/chain length/chain "difficulty"

→ Nonce
EpochNo nonce

→ NatNonce
Block nonce

→ UnitInterval
Praos leader value

→ Word
Period of KES (key evolving signature scheme)

→ Word
KES period of key registration

→ OCert (EraCrypto era)
Operational certificate

→ Block (BHeader (EraCrypto era)) era

Create a block with a faked VRF result.

mkOCert ∷ ∀ c (r ∷ KeyRole). (Crypto c, Signable (DSIGN c) (OCertSignable c)) ⇒ AllIssuerKeys c r → Word64 → KESPeriod → OCert c

getKESPeriodRenewalNo ∷ AllIssuerKeys h r → KESPeriod → Integer

Takes a sequence of KES hot keys and checks to see whether there is one whose range contains the current KES period. If so, return its index in the list of hot keys.

tooLateInEpoch ∷ SlotNo → Bool

True if the given slot is within the last `2 * stabilityWindow` slots of the current epoch.

data RawSeed

Constructors

RawSeed !Word64 !Word64 !Word64 !Word64 !Word64

mkKeyPair ∷ ∀ c kd. DSIGNAlgorithm (DSIGN c) ⇒ RawSeed → (SignKeyDSIGN (DSIGN c), VKey kd c)

For testing purposes, generate a deterministic key pair given a seed.

mkKeyPairs ∷ DSIGNAlgorithm (DSIGN c) ⇒ Word64 → (KeyPair kr c, KeyPair kr' c)

mkGenKey ∷ DSIGNAlgorithm (DSIGN c) ⇒ RawSeed → (SignKeyDSIGN (DSIGN c), VKey kd c)

For testing purposes, generate a deterministic genesis key pair given a seed.

genesisAccountState ∷ AccountState

Account with empty treasury

genCoin ∷ Integer → Integer → Gen Coin

type PreAlonzo era = TxWits era ~ ShelleyTxWits era

For use in the Serialisation and Example Tests, which assume Shelley, Allegra, or Mary Eras.

hashData ∷ ∀ era. Era era ⇒ Data → DataHash (EraCrypto era)

findPlutus ∷ ∀ era. Era era ⇒ GenEnv era → ScriptHash (EraCrypto era) → (Script era, StrictMaybe (DataHash (EraCrypto era)))

Find the preallocated Script from its Hash.