Test.Cardano.Ledger.Shelley.Generator.Delegation

Documentation

genDCert :: forall era. EraGen era => Constants -> KeySpace era -> PParams era -> AccountState -> DPState (Crypto era) -> SlotNo -> Gen (Maybe (DCert (Crypto era), CertCred era))

Occasionally generate a valid certificate

Returning Nothing indicates a failure to generate a value, usually due to lack of available values from the pre-populated (e.g. key) spaces. A Just represents a successfully generated value.

Different generators return witnesses that are either genesis or regular keys.

Note: we register keys and pools more often than deregistering/retiring them, and we generate more delegations than registrations of keys/pools.

data CertCred era

Constructors

CoreKeyCred [GenesisKeyPair (Crypto era)]

StakeCred (KeyPair 'Staking (Crypto era))

PoolCred (KeyPair 'StakePool (Crypto era))

ScriptCred (Script era, Script era)

DelegateCred [KeyPair 'GenesisDelegate (Crypto era)]

NoCred