Test.Cardano.Ledger.Shelley.Generator.Presets

Description

Pre-generated items to use in traces.

Functions in this module make specific assumptions about the sets of keys involved, and thus cannot be used as generic generators.

Documentation

coreNodeKeys ∷ Crypto c ⇒ Constants → [(KeyPair 'Genesis c, AllIssuerKeys c 'GenesisDelegate)]

keySpace ∷ ∀ era. EraGen era ⇒ Constants → KeySpace era

Example keyspace for use in generators

genEnv ∷ ∀ era. EraGen era ⇒ Proxy era → Constants → GenEnv era

Example generator environment, consisting of default constants and an corresponding keyspace.

genesisDelegs0 ∷ Crypto c ⇒ Constants → Map (KeyHash 'Genesis c) (GenDelegPair c)

someKeyPairs ∷ Crypto c ⇒ Constants → (Int, Int) → Gen (KeyPairs c)

Select between _lower_ and _upper_ keys from keyPairs

keyPairs ∷ Crypto c ⇒ Constants → KeyPairs c

Constant list of KeyPairs intended to be used in the generators.

scriptSpace ∷ ∀ era. EraScript era ⇒ [TwoPhase3ArgInfo era] → [TwoPhase2ArgInfo era] → ScriptSpace era

An Example Script space for use in Trace generators