Test.Cardano.Ledger.Shelley.Generator.ScriptClass

Documentation

class (UsesScript era, ValidateScript era, Crypto (Crypto era)) => ScriptClass era where

Minimal complete definition

basescript, isKey, quantify, unQuantify

Methods

basescript :: Proxy era -> KeyHash 'Witness (Crypto era) -> Script era

isKey :: Proxy era -> Script era -> Maybe (KeyHash 'Witness (Crypto era))

isOnePhase :: Proxy era -> Script era -> Bool

quantify :: Proxy era -> Script era -> Quantifier (Script era)

unQuantify :: Proxy era -> Quantifier (Script era) -> Script era

data Quantifier t

Constructors

AllOf [t]

AnyOf [t]

MOf Int [t]

Leaf t

exponential :: Integer -> Integer -> Gen Integer

anyOf :: forall era. ScriptClass era => Proxy era -> [Script era] -> Script era

allOf :: forall era. ScriptClass era => Proxy era -> [Script era] -> Script era

mOf :: forall era. ScriptClass era => Proxy era -> Int -> [Script era] -> Script era

keyPairs :: Crypto crypto => Constants -> KeyPairs crypto

Constant list of KeyPairs intended to be used in the generators.

mkPayScriptHashMap :: forall era. ScriptClass era => [(Script era, Script era)] -> Map (ScriptHash (Crypto era)) (Script era, Script era)

mkStakeScriptHashMap :: forall era. ScriptClass era => [(Script era, Script era)] -> Map (ScriptHash (Crypto era)) (Script era, Script era)

Generate a mapping from stake script hash to script pair.

mkScriptsFromKeyPair :: forall era. ScriptClass era => (KeyPair 'Payment (Crypto era), KeyPair 'Staking (Crypto era)) -> (Script era, Script era)

mkKeyPairs :: DSIGNAlgorithm (DSIGN crypto) => Word64 -> (KeyPair kr crypto, KeyPair kr' crypto)

mkScripts :: forall era. ScriptClass era => KeyPairs (Crypto era) -> [(Script era, Script era)]

make Scripts based on the given key pairs

mkScriptCombinations :: forall era. ScriptClass era => [(Script era, Script era)] -> [(Script era, Script era)]

Combine a list of script pairs into hierarchically structured multi-sig scripts, list must have at least length 3. Be careful not to call with too many pairs in order not to create too many of the possible combinations.

combinedScripts :: forall era. ScriptClass era => Constants -> [(Script era, Script era)]

Make a list of script pairs (payment,staking). Each of these are combined scripts. I.e. All, Any, MofN, etc. These come from combining the the first N (numBaseScripts) baseScripts When N==3, we get about 150 combined scripts.

baseScripts :: forall era. ScriptClass era => Constants -> [(Script era, Script era)]

Make list of script pairs (payment,staking). These are non-combined scripts Ie NO quantifer like All, Any, MofN, etc.) scripts. In post Shelley Eras, either Keylock or Require Start-Finish scripts.

scriptKeyCombinations :: forall era. ScriptClass era => Proxy era -> Script era -> [[KeyHash 'Witness (Crypto era)]]

Return all valid lists of KeyHashes that appear in a Script used in testing.

scriptKeyCombination :: forall era. ScriptClass era => Proxy era -> Script era -> [KeyHash 'Witness (Crypto era)]

Produce a valid list of key hashes that appear in a Script. Note: in the case of AnyOf, we just take the first script in the expression. This only works if we generate AnyOf scripts such that all script options are valid scripts (that is, valid in the context of a transaction, at generation time and execution/spend time).