Test.Cardano.Ledger.Shelley.Generator.Trace.Chain

Documentation

lastByronHeaderHash ∷ ∀ proxy era. proxy era → HashHeader

The first block of the Shelley era will point back to the last block of the Byron era. For our purposes we can bootstrap the chain by just coercing the value. When this transition actually occurs, the consensus layer will do the work of making sure that the hash gets translated across the fork

mkGenesisChainState ∷ ∀ era a c. (EraGen era, EraGov era, EraStake era) ⇒ GenEnv c era → IRC (CHAIN era) → Gen (Either a (ChainState era))

mkOCertIssueNos ∷ GenDelegs → Map (KeyHash 'BlockIssuer) Natural

registerGenesisStaking ∷ (EraGov era, EraStake era, EraCertState era, ShelleyEraAccounts era) ⇒ ShelleyGenesisStaking → ChainState era → ChainState era

Orphan instances

(EraGen era, EraBlockBody era, ApplyBlock era, GetLedgerView era, MinLEDGER_STS era, MinCHAIN_STS era, Embed (EraRule "BBODY" era) (CHAIN era), Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Signal (EraRule "BBODY" era) ~ Block BHeaderView era, Embed (EraRule "TICKN" era) (CHAIN era), Environment (EraRule "TICKN" era) ~ TicknEnv, State (EraRule "TICKN" era) ~ TicknState, Signal (EraRule "TICKN" era) ~ Bool, Embed (EraRule "TICK" era) (CHAIN era), Environment (EraRule "TICK" era) ~ (), State (EraRule "TICK" era) ~ NewEpochState era, Signal (EraRule "TICK" era) ~ SlotNo, HasTrace (EraRule "LEDGERS" era) (GenEnv MockCrypto era)) ⇒ HasTrace (CHAIN era) (GenEnv MockCrypto era)

Instance details

Associated Types

type BaseEnv (CHAIN era)

Methods

interpretSTS ∷ HasCallStack ⇒ BaseEnv (CHAIN era) → BaseM (CHAIN era) a → a

envGen ∷ GenEnv MockCrypto era → Gen (Environment (CHAIN era))

sigGen ∷ GenEnv MockCrypto era → Environment (CHAIN era) → State (CHAIN era) → Gen (Signal (CHAIN era))

shrinkSignal ∷ Signal (CHAIN era) → [Signal (CHAIN era)]