Test.Cardano.Ledger.Shelley.Generator.Trace.Ledger

Documentation

genAccountState ∷ Constants → Gen ChainAccountState

mkGenesisLedgerState ∷ ∀ a era ledger c. (EraGen era, EraGov era, EraStake era) ⇒ GenEnv c era → IRC ledger → Gen (Either a (LedgerState era))

Generate initial state for the LEDGER STS using the STS environment.

Note: this function must be usable in place of applySTS and needs to align with the signature 'RuleContext sts -> Gen (Either [[PredicateFailure sts]] (State sts))'. To achieve this we (1) use 'IRC LEDGER' (the "initial rule context") instead of simply LedgerEnv and (2) always return Right (since this function does not raise predicate failures).

Orphan instances

(EraGen era, EraGov era, EraUTxO era, EraCertState era, ShelleyEraAccounts era, MinLEDGER_STS era, Embed (EraRule "DELPL" era) (CERTS era), Environment (EraRule "DELPL" era) ~ DelplEnv era, State (EraRule "DELPL" era) ~ CertState era, Signal (EraRule "DELPL" era) ~ TxCert era, PredicateFailure (EraRule "DELPL" era) ~ ShelleyDelplPredFailure era, Embed (EraRule "DELEGS" era) (ShelleyLEDGER era), Embed (EraRule "UTXOW" era) (ShelleyLEDGER era), Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Signal (EraRule "UTXOW" era) ~ Tx era, Environment (EraRule "DELEGS" era) ~ DelegsEnv era, State (EraRule "DELEGS" era) ~ CertState era, Signal (EraRule "DELEGS" era) ~ Seq (TxCert era), AtMostEra "Babbage" era, Crypto c) ⇒ HasTrace (ShelleyLEDGER era) (GenEnv c era)

Instance details

Associated Types

type BaseEnv (ShelleyLEDGER era)

Methods

interpretSTS ∷ HasCallStack ⇒ BaseEnv (ShelleyLEDGER era) → BaseM (ShelleyLEDGER era) a → a

envGen ∷ GenEnv c era → Gen (Environment (ShelleyLEDGER era))

sigGen ∷ GenEnv c era → Environment (ShelleyLEDGER era) → State (ShelleyLEDGER era) → Gen (Signal (ShelleyLEDGER era))

shrinkSignal ∷ Signal (ShelleyLEDGER era) → [Signal (ShelleyLEDGER era)]

(Crypto c, EraGen era, EraGov era, EraUTxO era, EraStake era, EraCertState era, ShelleyEraAccounts era, MinLEDGER_STS era, Embed (EraRule "DELPL" era) (CERTS era), Environment (EraRule "DELPL" era) ~ DelplEnv era, State (EraRule "DELPL" era) ~ CertState era, Signal (EraRule "DELPL" era) ~ TxCert era, PredicateFailure (EraRule "DELPL" era) ~ ShelleyDelplPredFailure era, Embed (EraRule "DELEG" era) (ShelleyDELPL era), Embed (EraRule "LEDGER" era) (ShelleyLEDGERS era), AtMostEra "Babbage" era) ⇒ HasTrace (ShelleyLEDGERS era) (GenEnv c era)

Instance details

Associated Types

type BaseEnv (ShelleyLEDGERS era)

Methods

interpretSTS ∷ HasCallStack ⇒ BaseEnv (ShelleyLEDGERS era) → BaseM (ShelleyLEDGERS era) a → a

envGen ∷ GenEnv c era → Gen (Environment (ShelleyLEDGERS era))

sigGen ∷ GenEnv c era → Environment (ShelleyLEDGERS era) → State (ShelleyLEDGERS era) → Gen (Signal (ShelleyLEDGERS era))

shrinkSignal ∷ Signal (ShelleyLEDGERS era) → [Signal (ShelleyLEDGERS era)]