Test.Cardano.Ledger.Shelley.Generator.Trace.TxCert

Documentation

data CERTS era

This is a non-spec STS used to generate a sequence of certificates with witnesses.

genTxCerts ∷ ∀ era c. (EraGen era, Embed (EraRule "DELPL" era) (CERTS era), Environment (EraRule "DELPL" era) ~ DelplEnv era, State (EraRule "DELPL" era) ~ CertState era, Signal (EraRule "DELPL" era) ~ TxCert era, Crypto c) ⇒ GenEnv c era → PParams era → CertState era → SlotNo → TxIx → ChainAccountState → Gen ([TxCert era], Coin, Coin, CertState era, [KeyPair 'Witness], [(Script era, Script era)])

Generate certificates and also return the associated witnesses and deposits and refunds required.