Test.Cardano.Ledger.Shelley.Generator.TxCert

Documentation

genTxCert ∷ ∀ era. (EraGen era, ProtVerAtMost era 8) ⇒ Constants → KeySpace era → PParams era → AccountState → CertState era → SlotNo → Gen (Maybe (TxCert era, CertCred era))

Occasionally generate a valid certificate

Returning Nothing indicates a failure to generate a value, usually due to lack of available values from the pre-populated (e.g. key) spaces. A Just represents a successfully generated value.

Different generators return witnesses that are either genesis or regular keys.

Note: we register keys and pools more often than deregistering/retiring them, and we generate more delegations than registrations of keys/pools.

data CertCred era

Constructors

CoreKeyCred [GenesisKeyPair (EraCrypto era)]

StakeCred (KeyPair 'Staking (EraCrypto era))

PoolCred (KeyPair 'StakePool (EraCrypto era))

ScriptCred (Script era, Script era)

DelegateCred [KeyPair 'GenesisDelegate (EraCrypto era)]

NoCred