Test.Cardano.Ledger.Shelley.Generator.Update

Documentation

genPParams :: Constants -> Gen (PParams era)

genUpdate :: EraGen era => Constants -> SlotNo -> [(GenesisKeyPair (Crypto era), AllIssuerKeys (Crypto era) 'GenesisDelegate)] -> Map (KeyHash 'GenesisDelegate (Crypto era)) (AllIssuerKeys (Crypto era) 'GenesisDelegate) -> PParams era -> (UTxOState era, DPState (Crypto era)) -> Gen (Maybe (Update era), [KeyPair 'Witness (Crypto era)])

Occasionally generate an update and return with the witness keys

genShelleyPParamsDelta :: forall era. (PParams era ~ PParams era, PParamsDelta era ~ PParamsUpdate era) => Constants -> PParams era -> Gen (PParamsDelta era)

This is only good in the Shelley Era, used to define the genEraPParamsDelta method for (EraGen (ShelleyEra c))

genM :: Gen a -> Gen (StrictMaybe a)

genDecentralisationParam :: HasCallStack => Gen UnitInterval

^ TODO jc - generating d=0 takes some care, if there are no registered stake pools then d=0 deadlocks the system.