Test.Cardano.Ledger.Shelley.Rules.ClassifyTraces

Documentation

onlyValidLedgerSignalsAreGenerated :: forall era ledger. (EraGen era, ChainProperty era, HasTrace ledger (GenEnv era), Default (State (EraRule "PPUP" era)), BaseEnv ledger ~ Globals, State ledger ~ LedgerState era, Show (Environment ledger), Show (Signal ledger)) => Property

onlyValidChainSignalsAreGenerated :: forall era. (EraGen era, Default (State (EraRule "PPUP" era)), HasTrace (CHAIN era) (GenEnv era), Show (TxSeq era)) => Property

relevantCasesAreCovered :: forall era. (EraGen era, Default (State (EraRule "PPUP" era)), ChainProperty era, HasTrace (CHAIN era) (GenEnv era), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era)), HasField "update" (TxBody era) (StrictMaybe (Update era))) => Property

propAbstractSizeBoundsBytes :: forall era. (EraGen era, ChainProperty era, HasTrace (LEDGER era) (GenEnv era), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), ToCBOR (Tx era), Default (State (EraRule "PPUP" era))) => Property

Check that the abstract transaction size function actually bounds the number of bytes in the serialized transaction.

propAbstractSizeNotTooBig :: forall era. (EraGen era, ChainProperty era, HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), ToCBOR (Tx era), HasTrace (LEDGER era) (GenEnv era), Default (State (EraRule "PPUP" era))) => Property

Check that the abstract transaction size function is not off by an acceptable order of magnitude.