Test.Cardano.Ledger.Shelley.Rules.TestChain

Documentation

delegTraceFromBlock ∷ ∀ era. (ChainProperty era, ShelleyEraTxBody era, ShelleyEraAccounts era) ⇒ ChainState era → Block (BHeader MockCrypto) era → (DelegEnv era, Trace (ShelleyDELEG era))

Reconstruct a DELEG trace from all the transaction certificates in a Block

forAllChainTrace ∷ ∀ era prop. (EraGen era, EraGov era, EraStake era, Testable prop, HasTrace (CHAIN era) (GenEnv MockCrypto era)) ⇒ Word64 → Constants → (Trace (CHAIN era) → prop) → Property

ledgerTraceFromBlock ∷ ∀ era. (ChainProperty era, STS (EraRule "LEDGER" era), BaseM (EraRule "LEDGER" era) ~ ReaderT Globals Identity, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ LedgerState era, Signal (EraRule "LEDGER" era) ~ Tx era) ⇒ ChainState era → Block (BHeader MockCrypto) era → (ChainState era, Trace (EraRule "LEDGER" era))

Reconstruct a LEDGER trace from the transactions in a Block and ChainState

ledgerTraceFromBlockWithRestrictedUTxO ∷ ∀ era. (ChainProperty era, STS (EraRule "LEDGER" era), BaseM (EraRule "LEDGER" era) ~ ReaderT Globals Identity, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ LedgerState era, Signal (EraRule "LEDGER" era) ~ Tx era) ⇒ ChainState era → Block (BHeader MockCrypto) era → (UTxO era, Trace (EraRule "LEDGER" era))

This function is nearly the same as ledgerTraceFromBlock, but it restricts the UTxO state to only those needed by the block. It also returns the unused UTxO for comparison later.

chainSstWithTick ∷ ∀ era. ChainProperty era ⇒ Trace (CHAIN era) → [SourceSignalTarget (CHAIN era)]

Transform the [(source, signal, target)] of a CHAIN Trace by manually applying the Chain TICK Rule to each source and producing [(source, signal, target')].

This allows for testing properties on Chain traces while excluding the effects of Transactions and Certificates. For example we can check that pools that are due for retirement at an epoch transition, are indeed retired.

Had we not excluded the effects of Transactions/Certificates, we might have a pool that was correctly retired, but is again registered by a certificate in the block following the transition.

poolTraceFromBlock ∷ ∀ era. ChainProperty era ⇒ ChainState era → Block (BHeader MockCrypto) era → (ChainState era, Trace (ShelleyPOOL era))

Reconstruct a POOL trace from the transactions in a Block and ChainState

type TestingLedger era ledger = (BaseM ledger ~ ReaderT Globals Identity, Environment ledger ~ LedgerEnv era, State ledger ~ LedgerState era, Embed (EraRule "DELEGS" era) ledger, Embed (EraRule "UTXOW" era) ledger, STS ledger)

splitTrace ∷ (State s → State s → Bool) → Trace s → [Trace s]

Split a Trace into several shorter traces. The leftmost Trace (at the front of the list) might be ill-formed, depending on what p does.

forEachEpochTrace ∷ ∀ era prop. (EraGen era, EraGov era, EraStake era, Testable prop, HasTrace (CHAIN era) (GenEnv MockCrypto era)) ⇒ Int → Word64 → Constants → (Trace (CHAIN era) → prop) → Property

Test a property on the first subtracecount sub-Traces that end on an EpochBoundary

traceLen ∷ Word64

longTraceLen ∷ Word64

shortChainTrace ∷ ∀ era. (EraGen era, EraGov era, EraStake era, HasTrace (CHAIN era) (GenEnv MockCrypto era)) ⇒ Constants → (SourceSignalTarget (CHAIN era) → Property) → Property

Properties on really short chains, with only 100 successes