Test.Cardano.Ledger.Shelley.Rules.TestDeleg

Documentation

keyRegistration :: SourceSignalTarget (DELEG era) -> Property

Check stake key registration

keyDeRegistration :: SourceSignalTarget (DELEG era) -> Property

Check stake key de-registration

keyDelegation :: SourceSignalTarget (DELEG era) -> Property

Check stake key delegation

rewardsSumInvariant :: SourceSignalTarget (DELEG era) -> Property

Check that the sum of rewards does not change and that each element that is either removed or added has a zero balance.

checkInstantaneousRewards :: HasField "_protocolVersion" (PParams era) ProtVer => DelegEnv era -> SourceSignalTarget (DELEG era) -> Property