Test.Cardano.Ledger.Shelley.Serialisation.EraIndepGenerators

Documentation

mkDummyHash :: forall h a. HashAlgorithm h => Int -> Hash h a

genCoherentBlock :: forall era h. (ToCBORGroup (TxSeq era), Mock (Crypto era), UsesTxBody era, Arbitrary (Tx era), h ~ BHeader (Crypto era)) => Gen (Block h era)

For some purposes, a totally random block generator may not be suitable. There are tests in the ouroboros-network repository, for instance, that perform some integrity checks on the generated blocks.

For other purposes, such as the serialization tests in this repository, genBlock is more appropriate.

This generator uses mkBlock provide more coherent blocks.

genHash :: forall a h. HashAlgorithm h => Gen (Hash h a)

genShelleyAddress :: Crypto crypto => Gen (Addr crypto)

genByronAddress :: Gen (Addr crypto)

type MockGen era = (Mock (Crypto era), Arbitrary (VerKeyDSIGN (DSIGN (Crypto era))))

maxTxWits :: Int

Orphan instances

Arbitrary Natural

Instance details

Methods

arbitrary :: Gen Natural

shrink :: Natural -> [Natural]

Arbitrary Ptr

Instance details

Methods

arbitrary :: Gen Ptr

shrink :: Ptr -> [Ptr]

Arbitrary ProtVer

Instance details

Methods

arbitrary :: Gen ProtVer

shrink :: ProtVer -> [ProtVer]

Arbitrary NonNegativeInterval

Instance details

Methods

arbitrary :: Gen NonNegativeInterval

shrink :: NonNegativeInterval -> [NonNegativeInterval]

Arbitrary PositiveInterval

Instance details

Methods

arbitrary :: Gen PositiveInterval

shrink :: PositiveInterval -> [PositiveInterval]

Arbitrary PositiveUnitInterval

Instance details

Methods

arbitrary :: Gen PositiveUnitInterval

shrink :: PositiveUnitInterval -> [PositiveUnitInterval]

Arbitrary UnitInterval

Instance details

Methods

arbitrary :: Gen UnitInterval

shrink :: UnitInterval -> [UnitInterval]

Arbitrary Nonce

Instance details

Methods

arbitrary :: Gen Nonce

shrink :: Nonce -> [Nonce]

Arbitrary Url

Instance details

Methods

arbitrary :: Gen Url

shrink :: Url -> [Url]

Arbitrary DnsName

Instance details

Methods

arbitrary :: Gen DnsName

shrink :: DnsName -> [DnsName]

Arbitrary Port

Instance details

Methods

arbitrary :: Gen Port

shrink :: Port -> [Port]

Arbitrary ActiveSlotCoeff

Instance details

Methods

arbitrary :: Gen ActiveSlotCoeff

shrink :: ActiveSlotCoeff -> [ActiveSlotCoeff]

Arbitrary Network

Instance details

Methods

arbitrary :: Gen Network

shrink :: Network -> [Network]

Arbitrary TxIx

Instance details

Methods

arbitrary :: Gen TxIx

shrink :: TxIx -> [TxIx]

Arbitrary CertIx

Instance details

Methods

arbitrary :: Gen CertIx

shrink :: CertIx -> [CertIx]

Arbitrary Coin

Instance details

Methods

arbitrary :: Gen Coin

shrink :: Coin -> [Coin]

Arbitrary DeltaCoin

Instance details

Methods

arbitrary :: Gen DeltaCoin

shrink :: DeltaCoin -> [DeltaCoin]

Arbitrary EpochNo

Instance details

Methods

arbitrary :: Gen EpochNo

shrink :: EpochNo -> [EpochNo]

Arbitrary SlotNo

Instance details

Methods

arbitrary :: Gen SlotNo

shrink :: SlotNo -> [SlotNo]

Arbitrary VotingPeriod

Instance details

Methods

arbitrary :: Gen VotingPeriod

shrink :: VotingPeriod -> [VotingPeriod]

Arbitrary AccountState

Instance details

Methods

arbitrary :: Gen AccountState

shrink :: AccountState -> [AccountState]

Arbitrary LogWeight

Instance details

Methods

arbitrary :: Gen LogWeight

shrink :: LogWeight -> [LogWeight]

Arbitrary Likelihood

Instance details

Methods

arbitrary :: Gen Likelihood

shrink :: Likelihood -> [Likelihood]

Arbitrary PerformanceEstimate

Instance details

Methods

arbitrary :: Gen PerformanceEstimate

shrink :: PerformanceEstimate -> [PerformanceEstimate]

Arbitrary RewardType

Instance details

Methods

arbitrary :: Gen RewardType

shrink :: RewardType -> [RewardType]

Arbitrary Desirability

Instance details

Methods

arbitrary :: Gen Desirability

shrink :: Desirability -> [Desirability]

Arbitrary PoolMetadata

Instance details

Methods

arbitrary :: Gen PoolMetadata

shrink :: PoolMetadata -> [PoolMetadata]

Arbitrary StakePoolRelay

Instance details

Methods

arbitrary :: Gen StakePoolRelay

shrink :: StakePoolRelay -> [StakePoolRelay]

Arbitrary MIRPot

Instance details

Methods

arbitrary :: Gen MIRPot

shrink :: MIRPot -> [MIRPot]

Arbitrary Metadatum

Instance details

Methods

arbitrary :: Gen Metadatum

shrink :: Metadatum -> [Metadatum]

Arbitrary TicknState

Instance details

Methods

arbitrary :: Gen TicknState

shrink :: TicknState -> [TicknState]

Arbitrary IPv4

Instance details

Methods

arbitrary :: Gen IPv4

shrink :: IPv4 -> [IPv4]

Arbitrary IPv6

Instance details

Methods

arbitrary :: Gen IPv6

shrink :: IPv6 -> [IPv6]

DSIGNAlgorithm crypto => Arbitrary (VerKeyDSIGN crypto)

Instance details

Methods

arbitrary :: Gen (VerKeyDSIGN crypto)

shrink :: VerKeyDSIGN crypto -> [VerKeyDSIGN crypto]

Crypto crypto => Arbitrary (AuxiliaryDataHash crypto)

Instance details

Methods

arbitrary :: Gen (AuxiliaryDataHash crypto)

shrink :: AuxiliaryDataHash crypto -> [AuxiliaryDataHash crypto]

Arbitrary a => Arbitrary (StrictMaybe a)

Instance details

Methods

arbitrary :: Gen (StrictMaybe a)

shrink :: StrictMaybe a -> [StrictMaybe a]

Crypto crypto => Arbitrary (UnifiedMap crypto)

Instance details

Methods

arbitrary :: Gen (UnifiedMap crypto)

shrink :: UnifiedMap crypto -> [UnifiedMap crypto]

Crypto crypto => Arbitrary (Triple crypto)

Instance details

Methods

arbitrary :: Gen (Triple crypto)

shrink :: Triple crypto -> [Triple crypto]

Crypto crypto => Arbitrary (TxId crypto)

Instance details

Methods

arbitrary :: Gen (TxId crypto)

shrink :: TxId crypto -> [TxId crypto]

Crypto crypto => Arbitrary (TxIn crypto)

Instance details

Methods

arbitrary :: Gen (TxIn crypto)

shrink :: TxIn crypto -> [TxIn crypto]

Crypto crypto => Arbitrary (Addr crypto)

Instance details

Methods

arbitrary :: Gen (Addr crypto)

shrink :: Addr crypto -> [Addr crypto]

Crypto crypto => Arbitrary (RewardAcnt crypto)

Instance details

Methods

arbitrary :: Gen (RewardAcnt crypto)

shrink :: RewardAcnt crypto -> [RewardAcnt crypto]

Crypto crypto => Arbitrary (StakeReference crypto)

Instance details

Methods

arbitrary :: Gen (StakeReference crypto)

shrink :: StakeReference crypto -> [StakeReference crypto]

Crypto crypto => Arbitrary (BlocksMade crypto)

Instance details

Methods

arbitrary :: Gen (BlocksMade crypto)

shrink :: BlocksMade crypto -> [BlocksMade crypto]

Crypto crypto => Arbitrary (PoolDistr crypto)

Instance details

Methods

arbitrary :: Gen (PoolDistr crypto)

shrink :: PoolDistr crypto -> [PoolDistr crypto]

Crypto crypto => Arbitrary (GenDelegPair crypto)

Instance details

Methods

arbitrary :: Gen (GenDelegPair crypto)

shrink :: GenDelegPair crypto -> [GenDelegPair crypto]

Crypto crypto => Arbitrary (GenDelegs crypto)

Instance details

Methods

arbitrary :: Gen (GenDelegs crypto)

shrink :: GenDelegs crypto -> [GenDelegs crypto]

Crypto crypto => Arbitrary (ScriptHash crypto)

Instance details

Methods

arbitrary :: Gen (ScriptHash crypto)

shrink :: ScriptHash crypto -> [ScriptHash crypto]

Arbitrary (CompactForm Coin)

Instance details

Methods

arbitrary :: Gen (CompactForm Coin)

shrink :: CompactForm Coin -> [CompactForm Coin]

(Era era, Arbitrary (PredicateFailure (EraRule "LEDGER" era))) => Arbitrary (ApplyTxError era)

Instance details

Methods

arbitrary :: Gen (ApplyTxError era)

shrink :: ApplyTxError era -> [ApplyTxError era]

(Mock (Crypto era), Arbitrary (PParams era)) => Arbitrary (ShelleyGenesis era)

Instance details

Methods

arbitrary :: Gen (ShelleyGenesis era)

shrink :: ShelleyGenesis era -> [ShelleyGenesis era]

(Era era, Arbitrary (PredicateFailure (EraRule "LEDGER" era))) => Arbitrary (LedgersPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (LedgersPredicateFailure era)

shrink :: LedgersPredicateFailure era -> [LedgersPredicateFailure era]

(Era era, Arbitrary (PredicateFailure (EraRule "DELEGS" era)), Arbitrary (PredicateFailure (EraRule "UTXOW" era))) => Arbitrary (LedgerPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (LedgerPredicateFailure era)

shrink :: LedgerPredicateFailure era -> [LedgerPredicateFailure era]

(Era era, Mock (Crypto era), Arbitrary (PredicateFailure (EraRule "DELPL" era))) => Arbitrary (DelegsPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (DelegsPredicateFailure era)

shrink :: DelegsPredicateFailure era -> [DelegsPredicateFailure era]

(Era era, Mock (Crypto era), Arbitrary (PredicateFailure (EraRule "POOL" era)), Arbitrary (PredicateFailure (EraRule "DELEG" era))) => Arbitrary (DelplPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (DelplPredicateFailure era)

shrink :: DelplPredicateFailure era -> [DelplPredicateFailure era]

(Era era, Mock (Crypto era)) => Arbitrary (DelegPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (DelegPredicateFailure era)

shrink :: DelegPredicateFailure era -> [DelegPredicateFailure era]

Era era => Arbitrary (PoolPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (PoolPredicateFailure era)

shrink :: PoolPredicateFailure era -> [PoolPredicateFailure era]

(Era era, Arbitrary (PredicateFailure (EraRule "UTXO" era))) => Arbitrary (UtxowPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (UtxowPredicateFailure era)

shrink :: UtxowPredicateFailure era -> [UtxowPredicateFailure era]

Era era => Arbitrary (PpupPredicateFailure era)

Instance details

Methods

arbitrary :: Gen (PpupPredicateFailure era)

shrink :: PpupPredicateFailure era -> [PpupPredicateFailure era]

Crypto crypto => Arbitrary (FutureGenDeleg crypto)

Instance details

Methods

arbitrary :: Gen (FutureGenDeleg crypto)

shrink :: FutureGenDeleg crypto -> [FutureGenDeleg crypto]

Crypto crypto => Arbitrary (InstantaneousRewards crypto)

Instance details

Methods

arbitrary :: Gen (InstantaneousRewards crypto)

shrink :: InstantaneousRewards crypto -> [InstantaneousRewards crypto]

Crypto crypto => Arbitrary (DState crypto)

Instance details

Methods

arbitrary :: Gen (DState crypto)

shrink :: DState crypto -> [DState crypto]

Crypto crypto => Arbitrary (PState crypto)

Instance details

Methods

arbitrary :: Gen (PState crypto)

shrink :: PState crypto -> [PState crypto]

Crypto crypto => Arbitrary (DPState crypto)

Instance details

Methods

arbitrary :: Gen (DPState crypto)

shrink :: DPState crypto -> [DPState crypto]

(UsesTxOut era, UsesValue era, Mock (Crypto era), Arbitrary (TxOut era), Arbitrary (Value era), Arbitrary (PParams era), Arbitrary (State (EraRule "PPUP" era))) => Arbitrary (EpochState era)

Instance details

Methods

arbitrary :: Gen (EpochState era)

shrink :: EpochState era -> [EpochState era]

(Era era, Mock (Crypto era)) => Arbitrary (PPUPState era)

Instance details

Methods

arbitrary :: Gen (PPUPState era)

shrink :: PPUPState era -> [PPUPState era]

Crypto c => Arbitrary (IncrementalStake c)

Instance details

Methods

arbitrary :: Gen (IncrementalStake c)

shrink :: IncrementalStake c -> [IncrementalStake c]

(UsesTxOut era, UsesValue era, Mock (Crypto era), Arbitrary (TxOut era), Arbitrary (State (EraRule "PPUP" era))) => Arbitrary (UTxOState era)

Instance details

Methods

arbitrary :: Gen (UTxOState era)

shrink :: UTxOState era -> [UTxOState era]

(UsesTxOut era, UsesValue era, Mock (Crypto era), Arbitrary (TxOut era), Arbitrary (Value era), Arbitrary (PParams era), Arbitrary (State (EraRule "PPUP" era)), Arbitrary (StashedAVVMAddresses era)) => Arbitrary (NewEpochState era)

Instance details

Methods

arbitrary :: Gen (NewEpochState era)

shrink :: NewEpochState era -> [NewEpochState era]

(UsesTxOut era, UsesValue era, Mock (Crypto era), Arbitrary (TxOut era), Arbitrary (State (EraRule "PPUP" era))) => Arbitrary (LedgerState era)

Instance details

Methods

arbitrary :: Gen (LedgerState era)

shrink :: LedgerState era -> [LedgerState era]

Crypto crypto => Arbitrary (WitHashes crypto)

Instance details

Methods

arbitrary :: Gen (WitHashes crypto)

shrink :: WitHashes crypto -> [WitHashes crypto]

(UsesTxOut era, UsesValue era, Mock (Crypto era), Arbitrary (TxOut era)) => Arbitrary (UTxO era)

Instance details

Methods

arbitrary :: Gen (UTxO era)

shrink :: UTxO era -> [UTxO era]

Mock crypto => Arbitrary (Pulser crypto)

Instance details

Methods

arbitrary :: Gen (Pulser crypto)

shrink :: Pulser crypto -> [Pulser crypto]

Crypto crypto => Arbitrary (RewardUpdate crypto)

Instance details

Methods

arbitrary :: Gen (RewardUpdate crypto)

shrink :: RewardUpdate crypto -> [RewardUpdate crypto]

Mock crypto => Arbitrary (RewardSnapShot crypto)

Instance details

Methods

arbitrary :: Gen (RewardSnapShot crypto)

shrink :: RewardSnapShot crypto -> [RewardSnapShot crypto]

Mock crypto => Arbitrary (FreeVars crypto)

Instance details

Methods

arbitrary :: Gen (FreeVars crypto)

shrink :: FreeVars crypto -> [FreeVars crypto]

Mock crypto => Arbitrary (PulsingRewUpdate crypto)

Instance details

Methods

arbitrary :: Gen (PulsingRewUpdate crypto)

shrink :: PulsingRewUpdate crypto -> [PulsingRewUpdate crypto]

Crypto crypto => Arbitrary (NonMyopic crypto)

Instance details

Methods

arbitrary :: Gen (NonMyopic crypto)

shrink :: NonMyopic crypto -> [NonMyopic crypto]

Crypto crypto => Arbitrary (Reward crypto)

Instance details

Methods

arbitrary :: Gen (Reward crypto)

shrink :: Reward crypto -> [Reward crypto]

Crypto crypto => Arbitrary (LeaderOnlyReward crypto)

Instance details

Methods

arbitrary :: Gen (LeaderOnlyReward crypto)

shrink :: LeaderOnlyReward crypto -> [LeaderOnlyReward crypto]

Mock crypto => Arbitrary (PoolRewardInfo crypto)

Instance details

Methods

arbitrary :: Gen (PoolRewardInfo crypto)

shrink :: PoolRewardInfo crypto -> [PoolRewardInfo crypto]

Crypto crypto => Arbitrary (Stake crypto)

Instance details

Methods

arbitrary :: Gen (Stake crypto)

shrink :: Stake crypto -> [Stake crypto]

Crypto crypto => Arbitrary (SnapShot crypto)

Instance details

Methods

arbitrary :: Gen (SnapShot crypto)

shrink :: SnapShot crypto -> [SnapShot crypto]

Crypto crypto => Arbitrary (SnapShots crypto)

Instance details

Methods

arbitrary :: Gen (SnapShots crypto)

shrink :: SnapShots crypto -> [SnapShots crypto]

Mock crypto => Arbitrary (RewardProvenancePool crypto)

Instance details

Methods

arbitrary :: Gen (RewardProvenancePool crypto)

shrink :: RewardProvenancePool crypto -> [RewardProvenancePool crypto]

Mock crypto => Arbitrary (RewardProvenance crypto)

Instance details

Methods

arbitrary :: Gen (RewardProvenance crypto)

shrink :: RewardProvenance crypto -> [RewardProvenance crypto]

(Era era, Arbitrary (TxBody era), Arbitrary (Value era), Arbitrary (AuxiliaryData era), Arbitrary (Script era), Arbitrary (Witnesses era), ToCBOR (AuxiliaryData era), ToCBOR (TxBody era), ToCBOR (Witnesses era)) => Arbitrary (Tx era)

Instance details

Methods

arbitrary :: Gen (Tx era)

shrink :: Tx era -> [Tx era]

(UsesScript era, Mock (Crypto era), ValidateScript era, Arbitrary (Script era)) => Arbitrary (WitnessSet era)

Instance details

Methods

arbitrary :: Gen (WitnessSet era)

shrink :: WitnessSet era -> [WitnessSet era]

Crypto crypto => Arbitrary (Delegation crypto)

Instance details

Methods

arbitrary :: Gen (Delegation crypto)

shrink :: Delegation crypto -> [Delegation crypto]

Crypto crypto => Arbitrary (PoolParams crypto)

Instance details

Methods

arbitrary :: Gen (PoolParams crypto)

shrink :: PoolParams crypto -> [PoolParams crypto]

Crypto crypto => Arbitrary (Wdrl crypto)

Instance details

Methods

arbitrary :: Gen (Wdrl crypto)

shrink :: Wdrl crypto -> [Wdrl crypto]

(UsesValue era, Mock (Crypto era), Arbitrary (Value era)) => Arbitrary (TxOut era)

Instance details

Methods

arbitrary :: Gen (TxOut era)

shrink :: TxOut era -> [TxOut era]

Crypto crypto => Arbitrary (DelegCert crypto)

Instance details

Methods

arbitrary :: Gen (DelegCert crypto)

shrink :: DelegCert crypto -> [DelegCert crypto]

Crypto crypto => Arbitrary (PoolCert crypto)

Instance details

Methods

arbitrary :: Gen (PoolCert crypto)

shrink :: PoolCert crypto -> [PoolCert crypto]

Crypto crypto => Arbitrary (GenesisDelegCert crypto)

Instance details

Methods

arbitrary :: Gen (GenesisDelegCert crypto)

shrink :: GenesisDelegCert crypto -> [GenesisDelegCert crypto]

Crypto crypto => Arbitrary (MIRTarget crypto)

Instance details

Methods

arbitrary :: Gen (MIRTarget crypto)

shrink :: MIRTarget crypto -> [MIRTarget crypto]

Crypto crypto => Arbitrary (MIRCert crypto)

Instance details

Methods

arbitrary :: Gen (MIRCert crypto)

shrink :: MIRCert crypto -> [MIRCert crypto]

Crypto crypto => Arbitrary (DCert crypto)

Instance details

Methods

arbitrary :: Gen (DCert crypto)

shrink :: DCert crypto -> [DCert crypto]

Crypto crypto => Arbitrary (MultiSig crypto)

Instance details

Methods

arbitrary :: Gen (MultiSig crypto)

shrink :: MultiSig crypto -> [MultiSig crypto]

(Era era, Mock (Crypto era)) => Arbitrary (Update era)

Instance details

Methods

arbitrary :: Gen (Update era)

shrink :: Update era -> [Update era]

(Era era, Mock (Crypto era)) => Arbitrary (ProposedPPUpdates era)

Instance details

Methods

arbitrary :: Gen (ProposedPPUpdates era)

shrink :: ProposedPPUpdates era -> [ProposedPPUpdates era]

Arbitrary (Metadata era)

Instance details

Methods

arbitrary :: Gen (Metadata era)

shrink :: Metadata era -> [Metadata era]

Crypto crypto => Arbitrary (BootstrapWitness crypto)

Instance details

Methods

arbitrary :: Gen (BootstrapWitness crypto)

shrink :: BootstrapWitness crypto -> [BootstrapWitness crypto]

Crypto crypto => Arbitrary (PrtclState crypto)

Instance details

Methods

arbitrary :: Gen (PrtclState crypto)

shrink :: PrtclState crypto -> [PrtclState crypto]

Crypto crypto => Arbitrary (OBftSlot crypto)

Instance details

Methods

arbitrary :: Gen (OBftSlot crypto)

shrink :: OBftSlot crypto -> [OBftSlot crypto]

Crypto crypto => Arbitrary (HashHeader crypto)

Instance details

Methods

arbitrary :: Gen (HashHeader crypto)

shrink :: HashHeader crypto -> [HashHeader crypto]

Mock crypto => Arbitrary (BHeader crypto)

Instance details

Methods

arbitrary :: Gen (BHeader crypto)

shrink :: BHeader crypto -> [BHeader crypto]

Arbitrary a => Arbitrary (StrictSeq a)

Instance details

Methods

arbitrary :: Gen (StrictSeq a)

shrink :: StrictSeq a -> [StrictSeq a]

DSIGNAlgorithm crypto => Arbitrary (SignedDSIGN crypto a)

Instance details

Methods

arbitrary :: Gen (SignedDSIGN crypto a)

shrink :: SignedDSIGN crypto a -> [SignedDSIGN crypto a]

HashAlgorithm h => Arbitrary (Hash h a)

Instance details

Methods

arbitrary :: Gen (Hash h a)

shrink :: Hash h a -> [Hash h a]

(UsesTxBody era, ToCBORGroup (TxSeq era), SupportsSegWit era, Mock (Crypto era), Arbitrary (Tx era), h ~ BHeader (Crypto era)) => Arbitrary (Block h era)

Instance details

Methods

arbitrary :: Gen (Block h era)

shrink :: Block h era -> [Block h era]

Crypto crypto => Arbitrary (Credential r crypto)

Instance details

Methods

arbitrary :: Gen (Credential r crypto)

shrink :: Credential r crypto -> [Credential r crypto]

Arbitrary (VerKeyDSIGN (DSIGN crypto)) => Arbitrary (VKey kd crypto)

Instance details

Methods

arbitrary :: Gen (VKey kd crypto)

shrink :: VKey kd crypto -> [VKey kd crypto]

Crypto crypto => Arbitrary (KeyHash a crypto)

Instance details

Methods

arbitrary :: Gen (KeyHash a crypto)

shrink :: KeyHash a crypto -> [KeyHash a crypto]

HasAlgorithm c => Arbitrary (SafeHash c i)

Instance details

Methods

arbitrary :: Gen (SafeHash c i)

shrink :: SafeHash c i -> [SafeHash c i]

(Typeable kr, Crypto crypto) => Arbitrary (WitVKey kr crypto)

Instance details

Methods

arbitrary :: Gen (WitVKey kr crypto)

shrink :: WitVKey kr crypto -> [WitVKey kr crypto]

(Arbitrary k, Arbitrary v, Ord k, Vector kv k, Vector vv v) => Arbitrary (VMap kv vv k v)

Instance details

Methods

arbitrary :: Gen (VMap kv vv k v)

shrink :: VMap kv vv k v -> [VMap kv vv k v]