Test.Cardano.Ledger.Shelley.Serialisation.Generators

Orphan instances

Mock c => Arbitrary (UtxoPredicateFailure (ShelleyEra c))

Instance details

Methods

arbitrary :: Gen (UtxoPredicateFailure (ShelleyEra c))

shrink :: UtxoPredicateFailure (ShelleyEra c) -> [UtxoPredicateFailure (ShelleyEra c)]

Mock c => Arbitrary (TxBody (ShelleyEra c))

Instance details

Methods

arbitrary :: Gen (TxBody (ShelleyEra c))

shrink :: TxBody (ShelleyEra c) -> [TxBody (ShelleyEra c)]

Arbitrary (PParams era)
Note that this instance is a little off - it is an era-independent generator for something which is only valid in certain eras. Its sole use is for ShelleyGenesis, a somewhat confusing type which is in fact used as the genesis for multiple eras.

Instance details

Methods

arbitrary :: Gen (PParams era)

shrink :: PParams era -> [PParams era]