Test.Cardano.Ledger.Shelley.Serialisation.Generators.Genesis

Documentation

genShelleyGenesis :: Era era => Gen (ShelleyGenesis era)

genStaking :: Crypto crypto => Gen (ShelleyGenesisStaking crypto)

genPools :: Crypto crypto => Gen [(KeyHash 'StakePool crypto, PoolParams crypto)]

genStake :: Crypto crypto => Gen [(KeyHash 'Staking crypto, KeyHash 'StakePool crypto)]

genPoolParams :: forall crypto. Crypto crypto => Gen (PoolParams crypto)

genStakePoolRelay :: Gen StakePoolRelay

genStrictMaybe :: Gen a -> Gen (StrictMaybe a)

genDnsName :: Gen DnsName

genPoolMetadata :: Gen PoolMetadata

genPort :: Gen Port

genUrl :: Gen Url

genRewardAcnt :: Crypto crypto => Gen (RewardAcnt crypto)

genCredential :: Crypto crypto => Gen (Credential 'Staking crypto)

genHash :: HashAlgorithm v => Gen (Hash v a)

genWords :: Natural -> Gen [Word8]

genPParams :: Gen (PParams era)

genNatural :: Range Natural -> Gen Natural

genRational :: Gen Rational

genEpochNo :: Gen EpochNo

genMinUTxOValue :: Gen Coin

genNonce :: Gen Nonce

genProtVer :: Gen ProtVer

genUnitInterval :: Gen UnitInterval

genPositiveUnitInterval :: Gen PositiveUnitInterval

genNonNegativeInterval :: Gen NonNegativeInterval

genDecimalBoundedRational :: (Integral a, BoundedRational r) => (a -> Gen a) -> Gen r

Only numbers in Scientific format can roundtrip JSON, so we generate numbers that can be represented in both decimal form and some bounded type rational form.

genGenesisDelegationList :: Crypto crypto => Gen [(KeyHash 'Genesis crypto, GenDelegPair crypto)]

genGenesisDelegationPair :: forall crypto. Crypto crypto => Gen (KeyHash 'Genesis crypto, GenDelegPair crypto)

genVRFKeyHash :: forall crypto. Crypto crypto => Gen (Hash crypto (VerKeyVRF (VRF crypto)))

genVRFKeyPair :: forall crypto. Crypto crypto => Gen (SignKeyVRF (VRF crypto), VerKeyVRF (VRF crypto))

genFundsList :: Crypto crypto => Gen [(Addr crypto, Coin)]

genSeed :: Int -> Gen Seed

genKeyHash :: Crypto crypto => Gen (KeyHash disc crypto)

genKeyPair :: forall crypto krole. DSIGNAlgorithm (DSIGN crypto) => Gen (SignKeyDSIGN (DSIGN crypto), VKey krole crypto)

Generate a deterministic key pair given a seed.

genGenesisFundPair :: Crypto crypto => Gen (Addr crypto, Coin)

genAddress :: Crypto crypto => Gen (Addr crypto)

genNetworkMagic :: Gen Word32

genCoin :: Gen Coin

genSecurityParam :: Gen Word64

genSlotLength :: Gen NominalDiffTime

genUTCTime :: Gen UTCTime

genIPv4 :: Gen IPv4

genIPv6 :: Gen IPv6