Test.Cardano.Ledger.Shelley.Examples.Consensus

Documentation

type KeyPairWits era = [KeyPair 'Witness (EraCrypto era)]

data ShelleyResultExamples era

Constructors

ShelleyResultExamples

Fields

srePParams ∷ PParams era

sreProposedPPUpdates ∷ ProposedPPUpdates era

srePoolDistr ∷ PoolDistr (EraCrypto era)

sreNonMyopicRewards ∷ Map (Either Coin (Credential 'Staking (EraCrypto era))) (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

sreShelleyGenesis ∷ ShelleyGenesis (EraCrypto era)

data ShelleyLedgerExamples era

Constructors

ShelleyLedgerExamples

Fields

sleBlock ∷ Block (BHeader (EraCrypto era)) era

sleHashHeader ∷ HashHeader (EraCrypto era)

sleTx ∷ Tx era

sleApplyTxError ∷ ApplyTxError era

sleRewardsCredentials ∷ Set (Either Coin (Credential 'Staking (EraCrypto era)))

sleResultExamples ∷ ShelleyResultExamples era

sleNewEpochState ∷ NewEpochState era

sleChainDepState ∷ ChainDepState (EraCrypto era)

sleTranslationContext ∷ TranslationContext era

type ShelleyBasedEra' era = PraosCrypto (EraCrypto era)

defaultShelleyLedgerExamples ∷ ∀ era. (ShelleyBasedEra' era, EraSegWits era, EraGov era, PredicateFailure (EraRule "DELEGS" era) ~ ShelleyDelegsPredFailure era, PredicateFailure (EraRule "LEDGER" era) ~ ShelleyLedgerPredFailure era, Default (StashedAVVMAddresses era), ProtVerAtMost era 4) ⇒ (TxBody era → KeyPairWits era → TxWits era) → (ShelleyTx era → Tx era) → Value era → TxBody era → TxAuxData era → TranslationContext era → ShelleyLedgerExamples era

exampleShelleyLedgerBlock ∷ ∀ era. (EraSegWits era, PraosCrypto (EraCrypto era)) ⇒ Tx era → Block (BHeader (EraCrypto era)) era

exampleHashHeader ∷ ∀ era. ShelleyBasedEra' era ⇒ Proxy era → HashHeader (EraCrypto era)

mkKeyHash ∷ ∀ c discriminator. Crypto c ⇒ Int → KeyHash discriminator c

mkScriptHash ∷ ∀ c. Crypto c ⇒ Int → ScriptHash c

exampleTx ∷ ∀ era. EraTx era ⇒ (TxBody era → KeyPairWits era → TxWits era) → TxBody era → TxAuxData era → ShelleyTx era

This is not a valid transaction. We don't care, we are only interested in serialisation, not validation.

exampleProposedPParamsUpdates ∷ EraPParams era ⇒ ProposedPPUpdates era

examplePoolDistr ∷ ∀ c. PraosCrypto c ⇒ PoolDistr c

exampleNonMyopicRewards ∷ ∀ c. Crypto c ⇒ Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

testShelleyGenesis ∷ Crypto c ⇒ ShelleyGenesis c

These are dummy values.

exampleNewEpochState ∷ ∀ era. (EraTxOut era, EraGov era, ShelleyBasedEra' era, Default (StashedAVVMAddresses era)) ⇒ Value era → PParams era → PParams era → NewEpochState era

This is probably not a valid ledger. We don't care, we are only interested in serialisation, not validation.

exampleLedgerChainDepState ∷ ∀ c. Crypto c ⇒ Word64 → ChainDepState c

testEpochInfo ∷ EpochInfo Identity

mkDummyAnchor ∷ Crypto c ⇒ Int → Anchor c

ledgerExamplesShelley ∷ ShelleyLedgerExamples Shelley

ShelleyLedgerExamples for Shelley era

mkWitnessesPreAlonzo ∷ (EraTx era, Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody)) ⇒ Proxy era → TxBody era → KeyPairWits era → ShelleyTxWits era

exampleCoin ∷ Coin

exampleTxBodyShelley ∷ ShelleyTxBody Shelley

exampleAuxDataMap ∷ Map Word64 Metadatum

exampleAuxiliaryDataShelley ∷ TxAuxData Shelley

exampleTxIns ∷ Crypto c ⇒ Set (TxIn c)

exampleCerts ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ StrictSeq (TxCert era)

exampleWithdrawals ∷ Crypto c ⇒ Withdrawals c

exampleProposedPPUpdates ∷ EraPParams era ⇒ ProposedPPUpdates era

examplePayKey ∷ Crypto c ⇒ KeyPair 'Payment c

exampleStakeKey ∷ Crypto c ⇒ KeyPair 'Staking c

exampleKeys ∷ ∀ c r. Crypto c ⇒ AllIssuerKeys c r

keyToCredential ∷ Crypto c ⇒ KeyPair r c → Credential r c

mkDSIGNKeyPair ∷ ∀ c kd. DSIGNAlgorithm (DSIGN c) ⇒ Word8 → KeyPair kd c

mkKeyPair' from Test.Cardano.Ledger.Shelley.Utils doesn't work for real crypto: https://github.com/intersectmbo/cardano-ledger/issues/1770

mkVRFKeyPair ∷ ∀ c. Crypto c ⇒ Proxy c → Word8 → VRFKeyPair c

examplePoolParams ∷ ∀ c. Crypto c ⇒ PoolParams c