Test.Cardano.Ledger.Shelley.Examples.Federation

Description

The genesis/core nodes for Shelley Ledger Examples.

Documentation

numCoreNodes ∷ Word64

Number of Core Node

coreNodeSK ∷ ∀ c. Crypto c ⇒ Int → SignKeyDSIGN c

coreNodeVK ∷ ∀ c. Crypto c ⇒ Int → VKey 'Genesis c

Verification (Public) Keys

Retrieve the verification key for a core node by providing a number in the range [0, ... (numCoreNodes-1)].

coreNodeIssuerKeys ∷ ∀ c. Crypto c ⇒ Int → AllIssuerKeys c 'GenesisDelegate

Block Issuer Keys

Retrieve the block issuer keys (cold, VRF, and hot KES keys) for a core node by providing a number in the range [0, ... (numCoreNodes-1)].

coreNodeKeysBySchedule ∷ ∀ era. (HasCallStack, EraPParams era) ⇒ PParams era → Word64 → AllIssuerKeys (EraCrypto era) 'GenesisDelegate

Keys by Overlay Schedule

Retrieve all the keys associated with a core node for a given slot and protocol parameters. It will return an error if there is not a core node scheduled for the given slot.

genDelegs ∷ ∀ c. Crypto c ⇒ Map (KeyHash 'Genesis c) (GenDelegPair c)

Genesis Delegation Mapping

The map from genesis/core node (verification) key hashes to their delegate's (verification) key hash.