Test.Cardano.Ledger.Shelley.Generator.Core

Documentation

data AllIssuerKeys c (r ∷ KeyRole)

Constructors

AllIssuerKeys

Fields

aikCold ∷ KeyPair r

aikVrf ∷ VRFKeyPair c

aikHot ∷ NonEmpty (KESPeriod, KESKeyPair c)

aikColdKeyHash ∷ KeyHash r

data VRFKeyPair c

Constructors

VRFKeyPair

Fields

vrfSignKey ∷ !(SignKeyVRF (VRF c))

vrfVerKey ∷ !(VerKeyVRF (VRF c))

data KESKeyPair c

Constructors

KESKeyPair

Fields

kesSignKey ∷ !(UnsoundPureSignKeyKES (KES c))

kesVerKey ∷ !(VerKeyKES (KES c))

data GenEnv c era

Generator environment.

Constructors

GenEnv

Fields

geKeySpace ∷ KeySpace c era

geScriptSpapce ∷ ScriptSpace era

geConstants ∷ Constants

data ScriptSpace era

Constructors

ScriptSpace

Fields

ssScripts3 ∷ [TwoPhase3ArgInfo era]

A list of Two Phase 3 Arg Scripts and their associated data we can use.

ssScripts2 ∷ [TwoPhase2ArgInfo era]

A list of Two Phase 2 Arg Scripts and their associated data we can use.

ssHash3 ∷ Map ScriptHash (TwoPhase3ArgInfo era)

ssHash2 ∷ Map ScriptHash (TwoPhase2ArgInfo era)

data TwoPhase3ArgInfo era

Constructors

TwoPhase3ArgInfo

Fields

getScript3 ∷ Script era

A Plutus Script

getHash3 ∷ ScriptHash

Its ScriptHash

getData3 ∷ Data

A Data that will make it succeed

getRedeemer3 ∷ (Data, Natural, Natural)

A Redeemer that will make it succeed

getSucceeds3 ∷ Bool

data TwoPhase2ArgInfo era

Constructors

TwoPhase2ArgInfo

Fields

getScript2 ∷ Script era

A Plutus Script

getHash2 ∷ ScriptHash

Its ScriptHash

getRedeemer2 ∷ (Data, Natural, Natural)

A Redeemer that will make it succeed

getSucceeds2 ∷ Bool

type ScriptInfo era = (Map ScriptHash (TwoPhase3ArgInfo era), Map ScriptHash (TwoPhase2ArgInfo era))

data KeySpace c era

Collection of all keys which are required to generate a trace.

These are the _only_ keys which should be involved in the trace.

Constructors

KeySpace_

Fields

ksCoreNodes ∷ [(GenesisKeyPair c, AllIssuerKeys c 'GenesisDelegate)]

ksGenesisDelegates ∷ [AllIssuerKeys c 'GenesisDelegate]

Bag of keys to be used for future genesis delegates

ksStakePools ∷ [AllIssuerKeys c 'StakePool]

Bag of keys to be used for future stake pools

ksKeyPairs ∷ KeyPairs

Bag of keys to be used for future payment/staking addresses

ksMSigScripts ∷ [(Script era, Script era)]

ksIndexedPaymentKeys ∷ Map (KeyHash 'Payment) (KeyPair 'Payment)

Index over the payment keys in ksKeyPairs

ksIndexedStakingKeys ∷ Map (KeyHash 'Staking) (KeyPair 'Staking)

Index over the staking keys in ksKeyPairs

ksIndexedGenDelegates ∷ Map (KeyHash 'GenesisDelegate) (AllIssuerKeys c 'GenesisDelegate)

Index over the cold key hashes in Genesis Delegates

ksIndexedPayScripts ∷ Map ScriptHash (Script era, Script era)

Index over the pay script hashes in Script pairs

ksIndexedStakeScripts ∷ Map ScriptHash (Script era, Script era)

Index over the stake script hashes in Script pairs

pattern KeySpace ∷ ∀ c era. ScriptClass era ⇒ [(GenesisKeyPair c, AllIssuerKeys c 'GenesisDelegate)] → [AllIssuerKeys c 'GenesisDelegate] → [AllIssuerKeys c 'StakePool] → KeyPairs → [(Script era, Script era)] → KeySpace c era

newtype NatNonce

We provide our own nonces to mkBlock, which we then wish to recover as the output of the VRF functions. In general, however, we just derive them from a natural. Since the nonce is a hash, we do not want to recover it to find a preimage. In testing, therefore, we just wrap the raw natural, which we then encode into the fake VRF implementation.

Constructors

NatNonce Natural

findPayKeyPairAddr ∷ Addr → Map (KeyHash 'Payment) (KeyPair 'Payment) → KeyPair 'Payment

Find first matching key pair for address. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayKeyPairCred ∷ ∀ kr. Credential kr → Map (KeyHash kr) (KeyPair kr) → KeyPair kr

Find first matching key pair for a credential. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayScriptFromCred ∷ ∀ era. Credential 'Witness → Map ScriptHash (Script era, Script era) → (Script era, Script era)

Find matching multisig scripts for a credential.

findStakeScriptFromCred ∷ Credential 'Witness → Map ScriptHash (Script era, Script era) → (Script era, Script era)

Find first matching script for a credential.

findPayScriptFromAddr ∷ ∀ era. Addr → Map ScriptHash (Script era, Script era) → (Script era, Script era)

Find first matching multisig script for an address.

genBool ∷ Gen Bool

genCoinList ∷ Integer → Integer → Int → Gen [Coin]

Generates a list of Coin values of length between lower and upper and with values between minCoin and maxCoin.

genInteger ∷ Integer → Integer → Gen Integer

genNatural ∷ Natural → Natural → Gen Natural

Generator for a natural number between lower and upper

genWord64 ∷ Word64 → Word64 → Gen Word64

Generator for a Word64 between lower and upper

genTxOut ∷ ∀ era. EraTxOut era ⇒ Gen (Value era) → [Addr] → Gen [TxOut era]

Generates a list of coins for the given Addr and produced a TxOut for each Addr

Note: we need to keep the initial utxo coin sizes large enough so that when we simulate sequences of transactions, we have enough funds available to include certificates that require deposits.

genesisCoins ∷ TxId → [TxOut era] → UTxO era

Creates the UTxO for a new ledger with the specified genesis TxId and transaction outputs.

increasingProbabilityAt ∷ Gen a → (a, a) → Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

pickStakeKey ∷ KeyPairs → Gen (VKey 'Staking)

Select one random verification staking key from list of pairs of KeyPair.

mkAddr ∷ (MakeCredential p 'Payment, MakeStakeReference s) ⇒ p → s → Addr

Construct a Testnet address from payment and staking components

mkCredential ∷ MakeCredential c r ⇒ c → Credential r

mkBlock

Arguments

∷ ∀ era (r ∷ KeyRole) c. (Crypto c, EraBlockBody era, Signable (VRF c) Seed, Signable (KES c) (BHBody c))

⇒ HashHeader
Hash of previous block

→ AllIssuerKeys c r
All keys in the stake pool

→ [Tx era]
Transactions to record

→ SlotNo
Current slot

→ BlockNo
Block numberchain lengthchain "difficulty"

→ Nonce
EpochNo nonce

→ Word
Period of KES (key evolving signature scheme)

→ Word
KES period of key registration

→ OCert c
Operational certificate

→ Block (BHeader c) era

mkBlockFakeVRF

Arguments

∷ ∀ era (r ∷ KeyRole) c. (Crypto c, EraBlockBody era, Signable (VRF c) (WithResult Seed), Signable (KES c) (BHBody c))

⇒ HashHeader
Hash of previous block

→ AllIssuerKeys c r
All keys in the stake pool

→ [Tx era]
Transactions to record

→ SlotNo
Current slot

→ BlockNo
Block number/chain length/chain "difficulty"

→ Nonce
EpochNo nonce

→ NatNonce
Block nonce

→ UnitInterval
Praos leader value

→ Word
Period of KES (key evolving signature scheme)

→ Word
KES period of key registration

→ OCert c
Operational certificate

→ Block (BHeader c) era

Create a block with a faked VRF result.

mkOCert ∷ ∀ c (r ∷ KeyRole). Crypto c ⇒ AllIssuerKeys c r → Word64 → KESPeriod → OCert c

getKESPeriodRenewalNo ∷ AllIssuerKeys h r → KESPeriod → Integer

Takes a sequence of KES hot keys and checks to see whether there is one whose range contains the current KES period. If so, return its index in the list of hot keys.

tooLateInEpoch ∷ SlotNo → Bool

True if the given slot is within the last `2 * stabilityWindow` slots of the current epoch.

data RawSeed

Constructors

RawSeed !Word64 !Word64 !Word64 !Word64 !Word64

mkKeyPair ∷ ∀ kd. RawSeed → (SignKeyDSIGN DSIGN, VKey kd)

For testing purposes, generate a deterministic key pair given a seed.

mkKeyPairs ∷ Word64 → (KeyPair kr, KeyPair kr')

mkGenKey ∷ RawSeed → (SignKeyDSIGN DSIGN, VKey kd)

For testing purposes, generate a deterministic genesis key pair given a seed.

genesisChainAccountState ∷ ChainAccountState

Account with empty treasury

genCoin ∷ Integer → Integer → Gen Coin

type PreAlonzo era = TxWits era ~ ShelleyTxWits era

For use in the Serialisation and Example Tests, which assume Shelley, Allegra, or Mary Eras.

hashData ∷ Data → DataHash

findPlutus ∷ ∀ era c. GenEnv c era → ScriptHash → (Script era, StrictMaybe DataHash)

Find the preallocated Script from its Hash.