Test.Cardano.Ledger.Shelley.Generator.Core

Documentation

data AllIssuerKeys v (r :: KeyRole)

Constructors

AllIssuerKeys

Fields

cold :: KeyPair r v

vrf :: (SignKeyVRF v, VerKeyVRF v)

hot :: [(KESPeriod, (SignKeyKES v, VerKeyKES v))]

hk :: KeyHash r v

data GenEnv era

Generator environment.

Constructors

GenEnv

Fields

geKeySpace :: KeySpace era

geScriptSpapce :: ScriptSpace era

geConstants :: Constants

data ScriptSpace era

Constructors

ScriptSpace

Fields

ssScripts3 :: [TwoPhase3ArgInfo era]

A list of Two Phase 3 Arg Scripts and their associated data we can use.

ssScripts2 :: [TwoPhase2ArgInfo era]

A list of Two Phase 2 Arg Scripts and their associated data we can use.

ssHash3 :: Map (ScriptHash (Crypto era)) (TwoPhase3ArgInfo era)

ssHash2 :: Map (ScriptHash (Crypto era)) (TwoPhase2ArgInfo era)

data TwoPhase3ArgInfo era

Constructors

TwoPhase3ArgInfo

Fields

getScript3 :: Script era

A Plutus Script

getHash3 :: ScriptHash (Crypto era)

Its ScriptHash

getData3 :: Data

A Data that will make it succeed

getRedeemer3 :: (Data, Natural, Natural)

A Redeemer that will make it succeed

getSucceeds3 :: Bool

data TwoPhase2ArgInfo era

Constructors

TwoPhase2ArgInfo

Fields

getScript2 :: Script era

A Plutus Script

getHash2 :: ScriptHash (Crypto era)

Its ScriptHash

getRedeemer2 :: (Data, Natural, Natural)

A Redeemer that will make it succeed

getSucceeds2 :: Bool

type ScriptInfo era = (Map (ScriptHash (Crypto era)) (TwoPhase3ArgInfo era), Map (ScriptHash (Crypto era)) (TwoPhase2ArgInfo era))

data KeySpace era

Collection of all keys which are required to generate a trace.

These are the _only_ keys which should be involved in the trace.

Constructors

KeySpace_

Fields

ksCoreNodes :: [(GenesisKeyPair (Crypto era), AllIssuerKeys (Crypto era) 'GenesisDelegate)]

ksGenesisDelegates :: [AllIssuerKeys (Crypto era) 'GenesisDelegate]

Bag of keys to be used for future genesis delegates

ksStakePools :: [AllIssuerKeys (Crypto era) 'StakePool]

Bag of keys to be used for future stake pools

ksKeyPairs :: KeyPairs (Crypto era)

Bag of keys to be used for future payment/staking addresses

ksMSigScripts :: [(Script era, Script era)]

ksIndexedPaymentKeys :: Map (KeyHash 'Payment (Crypto era)) (KeyPair 'Payment (Crypto era))

Index over the payment keys in ksKeyPairs

ksIndexedStakingKeys :: Map (KeyHash 'Staking (Crypto era)) (KeyPair 'Staking (Crypto era))

Index over the staking keys in ksKeyPairs

ksIndexedGenDelegates :: Map (KeyHash 'GenesisDelegate (Crypto era)) (AllIssuerKeys (Crypto era) 'GenesisDelegate)

Index over the cold key hashes in Genesis Delegates

ksIndexedPayScripts :: Map (ScriptHash (Crypto era)) (Script era, Script era)

Index over the pay script hashes in Script pairs

ksIndexedStakeScripts :: Map (ScriptHash (Crypto era)) (Script era, Script era)

Index over the stake script hashes in Script pairs

pattern KeySpace :: forall era. ScriptClass era => [(GenesisKeyPair (Crypto era), AllIssuerKeys (Crypto era) 'GenesisDelegate)] -> [AllIssuerKeys (Crypto era) 'GenesisDelegate] -> [AllIssuerKeys (Crypto era) 'StakePool] -> KeyPairs (Crypto era) -> [(Script era, Script era)] -> KeySpace era

newtype NatNonce

We provide our own nonces to mkBlock, which we then wish to recover as the output of the VRF functions. In general, however, we just derive them from a natural. Since the nonce is a hash, we do not want to recover it to find a preimage. In testing, therefore, we just wrap the raw natural, which we then encode into the fake VRF implementation.

Constructors

NatNonce Natural

findPayKeyPairAddr :: forall era. Addr (Crypto era) -> Map (KeyHash 'Payment (Crypto era)) (KeyPair 'Payment (Crypto era)) -> KeyPair 'Payment (Crypto era)

Find first matching key pair for address. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayKeyPairCred :: forall era kr. Credential kr (Crypto era) -> Map (KeyHash kr (Crypto era)) (KeyPair kr (Crypto era)) -> KeyPair kr (Crypto era)

Find first matching key pair for a credential. Returns the matching key pair where the first element of the pair matched the hash in addr.

findPayScriptFromCred :: forall era. Credential 'Witness (Crypto era) -> Map (ScriptHash (Crypto era)) (Script era, Script era) -> (Script era, Script era)

Find matching multisig scripts for a credential.

findStakeScriptFromCred :: Credential 'Witness (Crypto era) -> Map (ScriptHash (Crypto era)) (Script era, Script era) -> (Script era, Script era)

Find first matching script for a credential.

findPayScriptFromAddr :: forall era. Addr (Crypto era) -> Map (ScriptHash (Crypto era)) (Script era, Script era) -> (Script era, Script era)

Find first matching multisig script for an address.

genBool :: Gen Bool

genCoinList :: Integer -> Integer -> Int -> Gen [Coin]

Generates a list of Coin values of length between lower and upper and with values between minCoin and maxCoin.

genInteger :: Integer -> Integer -> Gen Integer

genNatural :: Natural -> Natural -> Gen Natural

Generator for a natural number between lower and upper

genWord64 :: Word64 -> Word64 -> Gen Word64

Generator for a Word64 between lower and upper

genTxOut :: forall era. UsesTxOut era => Gen (Value era) -> [Addr (Crypto era)] -> Gen [TxOut era]

Generates a list of coins for the given Addr and produced a TxOut for each Addr

Note: we need to keep the initial utxo coin sizes large enough so that when we simulate sequences of transactions, we have enough funds available to include certificates that require deposits.

genesisCoins :: TxId (Crypto era) -> [TxOut era] -> UTxO era

Creates the UTxO for a new ledger with the specified genesis TxId and transaction outputs.

increasingProbabilityAt :: Gen a -> (a, a) -> Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

pickStakeKey :: KeyPairs crypto -> Gen (VKey 'Staking crypto)

Select one random verification staking key from list of pairs of KeyPair.

toAddr :: Crypto crypto => Network -> (KeyPair 'Payment crypto, KeyPair 'Staking crypto) -> Addr crypto

toCred :: forall crypto (kr :: KeyRole). Crypto crypto => KeyPair kr crypto -> Credential kr crypto

unitIntervalToNatural :: UnitInterval -> Natural

Try to map the unit interval to a natural number. We don't care whether this is surjective. But it should be right inverse to fromNatural - that is, one should be able to recover the UnitInterval value used here.

mkBlock

Arguments

:: forall era r. (UsesTxBody era, ToCBORGroup (TxSeq era), Mock (Crypto era))

=> HashHeader (Crypto era)
Hash of previous block

-> AllIssuerKeys (Crypto era) r
All keys in the stake pool

-> [Tx era]
Transactions to record

-> SlotNo
Current slot

-> BlockNo
Block numberchain lengthchain "difficulty"

-> Nonce
EpochNo nonce

-> Word
Period of KES (key evolving signature scheme)

-> Word
KES period of key registration

-> OCert (Crypto era)
Operational certificate

-> Block (BHeader (Crypto era)) era

mkBlockHeader

Arguments

:: Mock crypto

=> HashHeader crypto
Hash of previous block

-> AllIssuerKeys crypto r
All keys in the stake pool

-> SlotNo
Current slot

-> BlockNo
Block numberchain lengthchain "difficulty"

-> Nonce
EpochNo nonce

-> Word
Period of KES (key evolving signature scheme)

-> Word
KES period of key registration

-> OCert crypto
Operational certificate

-> Natural
Block size

-> Hash crypto EraIndependentBlockBody
Block body hash

-> BHeader crypto

mkBlockFakeVRF

Arguments

:: forall era r. (UsesTxBody era, ToCBORGroup (TxSeq era), ExMock (Crypto era))

=> HashHeader (Crypto era)
Hash of previous block

-> AllIssuerKeys (Crypto era) r
All keys in the stake pool

-> [Tx era]
Transactions to record

-> SlotNo
Current slot

-> BlockNo
Block numberchain lengthchain "difficulty"

-> Nonce
EpochNo nonce

-> NatNonce
Block nonce

-> UnitInterval
Praos leader value

-> Word
Period of KES (key evolving signature scheme)

-> Word
KES period of key registration

-> OCert (Crypto era)
Operational certificate

-> Block (BHeader (Crypto era)) era

Create a block with a faked VRF result.

mkOCert :: forall crypto r. (Crypto crypto, Signable (DSIGN crypto) (OCertSignable crypto)) => AllIssuerKeys crypto r -> Word64 -> KESPeriod -> OCert crypto

getKESPeriodRenewalNo :: AllIssuerKeys h r -> KESPeriod -> Integer

Takes a set of KES hot keys and checks to see whether there is one whose range contains the current KES period. If so, return its index in the list of hot keys.

tooLateInEpoch :: SlotNo -> Bool

True if the given slot is within the last `2 * stabilityWindow` slots of the current epoch.

data RawSeed

Constructors

RawSeed !Word64 !Word64 !Word64 !Word64 !Word64

mkKeyPair :: forall crypto kd. DSIGNAlgorithm (DSIGN crypto) => RawSeed -> (SignKeyDSIGN (DSIGN crypto), VKey kd crypto)

For testing purposes, generate a deterministic key pair given a seed.

mkKeyPairs :: DSIGNAlgorithm (DSIGN crypto) => Word64 -> (KeyPair kr crypto, KeyPair kr' crypto)

mkGenKey :: DSIGNAlgorithm (DSIGN crypto) => RawSeed -> (SignKeyDSIGN (DSIGN crypto), VKey kd crypto)

For testing purposes, generate a deterministic genesis key pair given a seed.

genesisAccountState :: AccountState

Account with empty treasury

genCoin :: Integer -> Integer -> Gen Coin

type PreAlonzo era = (Witnesses era ~ WitnessSet era, ToCBOR (AuxiliaryData era))

For use in the Serialisation and Example Tests, which assume Shelley, Allegra, or Mary Eras.

hashData :: forall era. Era era => Data -> DataHash (Crypto era)

findPlutus :: forall era. Era era => GenEnv era -> ScriptHash (Crypto era) -> (Script era, StrictMaybe (DataHash (Crypto era)))

Find the preallocated Script from its Hash.