Test.Cardano.Ledger.Shelley.Generator.Trace.Chain

Documentation

lastByronHeaderHash :: forall proxy era. Era era => proxy era -> HashHeader (Crypto era)

The first block of the Shelley era will point back to the last block of the Byron era. For our purposes we can bootstrap the chain by just coercing the value. When this transition actually occurs, the consensus layer will do the work of making sure that the hash gets translated across the fork

mkGenesisChainState :: forall era a. (Default (State (EraRule "PPUP" era)), EraGen era) => GenEnv era -> IRC (CHAIN era) -> Gen (Either a (ChainState era))

mkOCertIssueNos :: GenDelegs h -> Map (KeyHash 'BlockIssuer h) Natural

registerGenesisStaking :: forall era. ShelleyGenesisStaking (Crypto era) -> ChainState era -> ChainState era

Orphan instances

(Era era, EraGen era, UsesTxBody era, UsesTxOut era, UsesValue era, UsesAuxiliary era, Mock (Crypto era), ApplyBlock era, GetLedgerView era, MinLEDGER_STS era, MinCHAIN_STS era, Embed (EraRule "BBODY" era) (CHAIN era), Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ BbodyState era, Signal (EraRule "BBODY" era) ~ Block (BHeaderView (Crypto era)) era, Embed (EraRule "TICKN" era) (CHAIN era), Environment (EraRule "TICKN" era) ~ TicknEnv, State (EraRule "TICKN" era) ~ TicknState, Signal (EraRule "TICKN" era) ~ Bool, Embed (EraRule "TICK" era) (CHAIN era), Environment (EraRule "TICK" era) ~ (), State (EraRule "TICK" era) ~ NewEpochState era, Signal (EraRule "TICK" era) ~ SlotNo, HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "outputs" (TxBody era) (StrictSeq (TxOut era)), HasField "_d" (PParams era) UnitInterval, HasTrace (EraRule "LEDGERS" era) (GenEnv era)) => HasTrace (CHAIN era) (GenEnv era)

Instance details

Associated Types

type BaseEnv (CHAIN era)

Methods

interpretSTS :: HasCallStack => BaseEnv (CHAIN era) -> BaseM (CHAIN era) a -> a

envGen :: GenEnv era -> Gen (Environment (CHAIN era))

sigGen :: GenEnv era -> Environment (CHAIN era) -> State (CHAIN era) -> Gen (Signal (CHAIN era))

shrinkSignal :: Signal (CHAIN era) -> [Signal (CHAIN era)]