Test.Cardano.Ledger.Shelley.Generator.Trace.Ledger

Documentation

genAccountState :: Constants -> Gen AccountState

mkGenesisLedgerState :: forall a era ledger. (UsesValue era, EraGen era, Default (State (EraRule "PPUP" era))) => GenEnv era -> IRC ledger -> Gen (Either a (LedgerState era))

Generate initial state for the LEDGER STS using the STS environment.

Note: this function must be usable in place of applySTS and needs to align with the signature 'RuleContext sts -> Gen (Either [[PredicateFailure sts]] (State sts))'. To achieve this we (1) use 'IRC LEDGER' (the "initial rule context") instead of simply LedgerEnv and (2) always return Right (since this function does not raise predicate failures).

Orphan instances

(EraGen era, UsesTxBody era, UsesTxOut era, UsesValue era, UsesAuxiliary era, Mock (Crypto era), MinLEDGER_STS era, Embed (EraRule "DELPL" era) (CERTS era), Environment (EraRule "DELPL" era) ~ DelplEnv era, State (EraRule "DELPL" era) ~ DPState (Crypto era), Signal (EraRule "DELPL" era) ~ DCert (Crypto era), PredicateFailure (EraRule "DELPL" era) ~ DelplPredicateFailure era, Embed (EraRule "DELEG" era) (DELPL era), Embed (EraRule "LEDGER" era) (LEDGERS era), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "outputs" (TxBody era) (StrictSeq (TxOut era)), Default (State (EraRule "PPUP" era))) => HasTrace (LEDGERS era) (GenEnv era)

Instance details

Associated Types

type BaseEnv (LEDGERS era)

Methods

interpretSTS :: HasCallStack => BaseEnv (LEDGERS era) -> BaseM (LEDGERS era) a -> a

envGen :: GenEnv era -> Gen (Environment (LEDGERS era))

sigGen :: GenEnv era -> Environment (LEDGERS era) -> State (LEDGERS era) -> Gen (Signal (LEDGERS era))

shrinkSignal :: Signal (LEDGERS era) -> [Signal (LEDGERS era)]

(EraGen era, UsesTxBody era, UsesTxOut era, UsesValue era, UsesAuxiliary era, Mock (Crypto era), MinLEDGER_STS era, TransValue ToCBOR era, Embed (EraRule "DELPL" era) (CERTS era), Environment (EraRule "DELPL" era) ~ DelplEnv era, State (EraRule "DELPL" era) ~ DPState (Crypto era), Signal (EraRule "DELPL" era) ~ DCert (Crypto era), PredicateFailure (EraRule "DELPL" era) ~ DelplPredicateFailure era, Embed (EraRule "DELEGS" era) (LEDGER era), Embed (EraRule "UTXOW" era) (LEDGER era), Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Signal (EraRule "UTXOW" era) ~ Tx era, Environment (EraRule "DELEGS" era) ~ DelegsEnv era, State (EraRule "DELEGS" era) ~ DPState (Crypto era), Signal (EraRule "DELEGS" era) ~ Seq (DCert (Crypto era)), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "outputs" (TxBody era) (StrictSeq (TxOut era)), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), Show (State (EraRule "PPUP" era)), Show (Tx era)) => HasTrace (LEDGER era) (GenEnv era)

Instance details

Associated Types

type BaseEnv (LEDGER era)

Methods

interpretSTS :: HasCallStack => BaseEnv (LEDGER era) -> BaseM (LEDGER era) a -> a

envGen :: GenEnv era -> Gen (Environment (LEDGER era))

sigGen :: GenEnv era -> Environment (LEDGER era) -> State (LEDGER era) -> Gen (Signal (LEDGER era))

shrinkSignal :: Signal (LEDGER era) -> [Signal (LEDGER era)]