Test.Cardano.Ledger.Shelley.Generator.Update

Documentation

genPParams ∷ ∀ era. (EraPParams era, ProtVerAtMost era 4, ProtVerAtMost era 6) ⇒ Constants → Gen (PParams era)

genUpdate ∷ EraGen era ⇒ Constants → SlotNo → [(GenesisKeyPair (EraCrypto era), AllIssuerKeys (EraCrypto era) 'GenesisDelegate)] → Map (KeyHash 'GenesisDelegate (EraCrypto era)) (AllIssuerKeys (EraCrypto era) 'GenesisDelegate) → PParams era → (UTxOState era, CertState era) → Gen (Maybe (Update era), [KeyPair 'Witness (EraCrypto era)])

Occasionally generate an update and return with the witness keys

genShelleyPParamsUpdate ∷ ∀ era. (ProtVerAtMost era 4, ProtVerAtMost era 6, ProtVerAtMost era 8, EraPParams era) ⇒ Constants → PParams era → Gen (PParamsUpdate era)

This is only good in the Shelley Era, used to define the genShelleyEraPParamsUpdate method for (EraGen (ShelleyEra c))

genM ∷ Gen a → Gen (StrictMaybe a)

genDecentralisationParam ∷ HasCallStack ⇒ Gen UnitInterval

^ TODO jc - generating d=0 takes some care, if there are no registered stake pools then d=0 deadlocks the system.