Test.Cardano.Ledger.Shelley.Rewards

Description

Currently this uses the trace mechanism to check that computing rewards has a required set of properties. It works only in the Shelley Era. It could be generalized, and then moved to the GeneratorTrace directory which computes property tests in all eras.

Documentation

rewardTests :: TestTree

type C = ShelleyEra C_Crypto

defaultMain :: TestTree -> IO ()

Parse the command line arguments and run the tests.

When the tests finish, this function calls exitWith with the exit code that indicates whether any tests have failed. Most external systems (stack, cabal, travis-ci, jenkins etc.) rely on the exit code to detect whether the tests pass. If you want to do something else after defaultMain returns, you need to catch the exception and then re-throw it. Example:

import Test.Tasty
import Test.Tasty.HUnit
import System.Exit
import Control.Exception

test = testCase "Test 1" (2 @?= 3)

main = defaultMain test
  `catch` (\e -> do
    if e == ExitSuccess
      then putStrLn "Yea"
      else putStrLn "Nay"
    throwIO e)

newEpochProp :: Word64 -> (NewEpochState C -> Property) -> Property

Provide a legitimate NewEpochState to make an test Property

newEpochEventsProp :: Word64 -> ([ChainEvent C] -> NewEpochState C -> Property) -> Property

Given a NewEpochState and [ChainEvent], test a Property at every Epoch Boundary

ppAgg :: Map (Credential 'Staking (Crypto C)) (Set (Reward (Crypto C))) -> PDoc

data RewardUpdateOld crypto

Constructors

RewardUpdateOld

Fields

deltaTOld :: !DeltaCoin

deltaROld :: !DeltaCoin

rsOld :: !(Map (Credential 'Staking crypto) Coin)

deltaFOld :: !DeltaCoin

nonMyopicOld :: !(NonMyopic crypto)

createRUpdOld :: forall era. (HasField "_d" (PParams era) UnitInterval, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, HasField "_protocolVersion" (PParams era) ProtVer, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural) => EpochSize -> BlocksMade (Crypto era) -> EpochState era -> Coin -> ShelleyBase (RewardUpdateOld (Crypto era))

createRUpdOld_ :: forall era. (HasField "_d" (PParams era) UnitInterval, HasField "_rho" (PParams era) UnitInterval, HasField "_tau" (PParams era) UnitInterval, HasField "_protocolVersion" (PParams era) ProtVer, HasField "_a0" (PParams era) NonNegativeInterval, HasField "_nOpt" (PParams era) Natural) => EpochSize -> BlocksMade (Crypto era) -> SnapShots (Crypto era) -> Coin -> PParams era -> Coin -> Set (Credential 'Staking (Crypto era)) -> NonMyopic (Crypto era) -> ShelleyBase (RewardUpdateOld (Crypto era))

Orphan instances

PrettyA (Reward crypto)

Instance details

Methods

prettyA :: Reward crypto -> PDoc

PrettyA x => Terse (Set x)

Instance details

Methods

terse :: Set x -> String

Terse (Reward crypto)

Instance details

Methods

terse :: Reward crypto -> String