Test.Cardano.Ledger.Shelley.Rules.Chain

Documentation

data CHAIN era

data ChainState era

Constructors

ChainState

Fields

chainNes ∷ NewEpochState era

chainOCertIssue ∷ Map (KeyHash 'BlockIssuer) Word64

chainEpochNonce ∷ Nonce

chainEvolvingNonce ∷ Nonce

chainCandidateNonce ∷ Nonce

chainPrevEpochNonce ∷ Nonce

chainLastAppliedBlock ∷ WithOrigin LastAppliedBlock

data TestChainPredicateFailure era

Constructors

RealChainPredicateFailure ChainPredicateFailure

BbodyFailure (PredicateFailure (EraRule "BBODY" era))

TicknFailure (PredicateFailure (EraRule "TICKN" era))

PrtclFailure (PredicateFailure (PRTCL MockCrypto))

PrtclSeqFailure PrtlSeqFailure

data ChainEvent era

Constructors

BbodyEvent !(Event (EraRule "BBODY" era))

TickEvent !(Event (EraRule "TICK" era))

TicknEvent !(Event (EraRule "TICKN" era))

PrtclEvent !(Event (PRTCL MockCrypto))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data AdaPots

Constructors

AdaPots

Fields

treasuryAdaPot ∷ Coin

reservesAdaPot ∷ Coin

rewardsAdaPot ∷ Coin

utxoAdaPot ∷ Coin

feesAdaPot ∷ Coin

obligationsPot ∷ Obligations

initialShelleyState ∷ ∀ era. (EraGov era, EraStake era, EraCertState era, Default (StashedAVVMAddresses era)) ⇒ WithOrigin LastAppliedBlock → EpochNo → UTxO era → Coin → Map (KeyHash 'Genesis) GenDelegPair → PParams era → Nonce → ChainState era

Creates a valid initial chain state

totalAda ∷ (EraTxOut era, EraGov era, EraCertState era) ⇒ ChainState era → Coin

Calculate the total ada in the chain state

totalAdaPots ∷ (EraTxOut era, EraGov era, EraCertState era) ⇒ ChainState era → AdaPots

Calculate the total ada pots in the chain state

chainStateNesL ∷ Lens' (ChainState era) (NewEpochState era)

Orphan instances

ToExpr HashHeader

Instance details

Methods

toExpr ∷ HashHeader → Expr

listToExpr ∷ [HashHeader] → Expr

ToExpr LastAppliedBlock

Instance details

Methods

toExpr ∷ LastAppliedBlock → Expr

listToExpr ∷ [LastAppliedBlock] → Expr