Test.Cardano.Ledger.Shelley.Rules.Chain

Documentation

data CHAIN era

data ChainState era

Constructors

ChainState

Fields

chainNes ∷ NewEpochState era

chainOCertIssue ∷ Map (KeyHash 'BlockIssuer (EraCrypto era)) Word64

chainEpochNonce ∷ Nonce

chainEvolvingNonce ∷ Nonce

chainCandidateNonce ∷ Nonce

chainPrevEpochNonce ∷ Nonce

chainLastAppliedBlock ∷ WithOrigin (LastAppliedBlock (EraCrypto era))

data TestChainPredicateFailure era

Constructors

RealChainPredicateFailure !ChainPredicateFailure

BbodyFailure !(PredicateFailure (EraRule "BBODY" era))

TickFailure !(PredicateFailure (EraRule "TICK" era))

TicknFailure !(PredicateFailure (EraRule "TICKN" era))

PrtclFailure !(PredicateFailure (PRTCL (EraCrypto era)))

PrtclSeqFailure !(PrtlSeqFailure (EraCrypto era))

data ChainEvent era

Constructors

BbodyEvent !(Event (EraRule "BBODY" era))

TickEvent !(Event (EraRule "TICK" era))

TicknEvent !(Event (EraRule "TICKN" era))

PrtclEvent !(Event (PRTCL (EraCrypto era)))

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data AdaPots

Constructors

AdaPots

Fields

treasuryAdaPot ∷ Coin

reservesAdaPot ∷ Coin

rewardsAdaPot ∷ Coin

utxoAdaPot ∷ Coin

keyDepositAdaPot ∷ Coin

poolDepositAdaPot ∷ Coin

depositsAdaPot ∷ Coin

feesAdaPot ∷ Coin

initialShelleyState ∷ (EraTxOut era, EraGovernance era, Default (StashedAVVMAddresses era)) ⇒ WithOrigin (LastAppliedBlock (EraCrypto era)) → EpochNo → UTxO era → Coin → Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)) → PParams era → Nonce → ChainState era

Creates a valid initial chain state

totalAda ∷ EraTxOut era ⇒ ChainState era → Coin

Calculate the total ada in the chain state

totalAdaPots ∷ EraTxOut era ⇒ ChainState era → AdaPots

Calculate the total ada pots in the chain state

Orphan instances

ToExpr (HashHeader c)

Instance details

Methods

toExpr ∷ HashHeader c → Expr

listToExpr ∷ [HashHeader c] → Expr

ToExpr (LastAppliedBlock c)

Instance details

Methods

toExpr ∷ LastAppliedBlock c → Expr

listToExpr ∷ [LastAppliedBlock c] → Expr