Test.Cardano.Ledger.Shelley.Rules.ClassifyTraces

Documentation

onlyValidLedgerSignalsAreGenerated ∷ ∀ era ledger. (EraGen era, HasTrace ledger (GenEnv era), BaseEnv ledger ~ Globals, State ledger ~ LedgerState era, Show (Environment ledger), Show (Signal ledger), EraGov era) ⇒ TestTree

onlyValidChainSignalsAreGenerated ∷ ∀ era. (EraGen era, HasTrace (CHAIN era) (GenEnv era), EraGov era) ⇒ TestTree

relevantCasesAreCovered ∷ ∀ era. (EraGen era, ChainProperty era, HasTrace (CHAIN era) (GenEnv era)) ⇒ Int → TestTree

propAbstractSizeBoundsBytes ∷ ∀ era. (EraGen era, HasTrace (ShelleyLEDGER era) (GenEnv era), EraGov era) ⇒ Property

Check that the abstract transaction size function actually bounds the number of bytes in the serialized transaction.

propAbstractSizeNotTooBig ∷ ∀ era. (EraGen era, HasTrace (ShelleyLEDGER era) (GenEnv era), EraGov era) ⇒ Property

Check that the abstract transaction size function is not off by an acceptable order of magnitude.