Test.Cardano.Ledger.Shelley.Rules.IncrementalStake

Documentation

incrStakeComputationTest ∷ ∀ era ledger. (EraGen era, TestingLedger era ledger, ChainProperty era, HasTrace (CHAIN era) (GenEnv era)) ⇒ TestTree

incrStakeComparisonTest ∷ ∀ era. (EraGen era, HasTrace (CHAIN era) (GenEnv era), EraGov era) ⇒ Proxy era → TestTree

stakeDistr ∷ ∀ era. EraTxOut era ⇒ UTxO era → DState era → PState era → SnapShot (EraCrypto era)

Compute the current Stake Distribution. This was called at the Epoch boundary in the Snap Rule. Now it is called in the tests to see that its incremental analog incrementalStakeDistr agrees.

aggregateUtxoCoinByCredential ∷ ∀ era. EraTxOut era ⇒ Map Ptr (Credential 'Staking (EraCrypto era)) → UTxO era → Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) Coin

Sum up all the Coin for each staking Credential. This function has an incremental analog. See incrementalAggregateUtxoCoinByCredential