Test.Cardano.Ledger.Shelley.Rules.TestChain

Documentation

removedAfterPoolreap :: forall era. (ChainProperty era, Default (State (EraRule "PPUP" era)), EraGen era, HasTrace (CHAIN era) (GenEnv era)) => Property

adaPreservationChain :: forall era ledger. (EraGen era, TestingLedger era ledger, State (EraRule "PPUP" era) ~ PPUPState era, ChainProperty era, HasTrace (CHAIN era) (GenEnv era), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era))) => Property

Various preservation properties

collisionFreeComplete :: forall era ledger. (EraGen era, ChainProperty era, TestingLedger era ledger, Default (State (EraRule "PPUP" era)), HasTrace (CHAIN era) (GenEnv era), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "addrWits" (Witnesses era) (Set (WitVKey 'Witness (Crypto era))), HasField "scriptWits" (Witnesses era) (Map (ScriptHash (Crypto era)) (Script era))) => Property

Tx inputs are eliminated, outputs added to utxo and TxIds are unique

poolProperties :: forall era. (EraGen era, Default (State (EraRule "PPUP" era)), ChainProperty era, HasTrace (CHAIN era) (GenEnv era), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era)))) => Property

Various properties of the POOL STS Rule, tested on longer traces (double the default length)

delegProperties :: forall era. (EraGen era, Default (State (EraRule "PPUP" era)), HasTrace (CHAIN era) (GenEnv era), ChainProperty era, HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era)))) => Property

Various properties of the POOL STS Rule, tested on longer traces (double the default length)

forAllChainTrace :: forall era prop. (Testable prop, Default (State (EraRule "PPUP" era)), EraGen era, HasTrace (CHAIN era) (GenEnv era)) => Word64 -> (Trace (CHAIN era) -> prop) -> Property

ledgerTraceFromBlock :: forall era ledger. (ChainProperty era, TestingLedger era ledger) => ChainState era -> Block (BHeader (Crypto era)) era -> (ChainState era, Trace ledger)

Reconstruct a LEDGER trace from the transactions in a Block and ChainState

type TestingLedger era ledger = (BaseM ledger ~ ReaderT Globals Identity, Environment ledger ~ LedgerEnv era, State ledger ~ LedgerState era, Signal ledger ~ Tx era, Embed (EraRule "DELEGS" era) ledger, Embed (EraRule "UTXOW" era) ledger, STS ledger)

stakeDistr :: forall era. Era era => UTxO era -> DState (Crypto era) -> PState (Crypto era) -> SnapShot (Crypto era)

Compute the current Stake Distribution. This was called at the Epoch boundary in the Snap Rule. Now it is called in the tests to see that its incremental analog incrementalStakeDistr agrees.

stakeIncrTest :: forall era ledger. (EraGen era, TestingLedger era ledger, State (EraRule "PPUP" era) ~ PPUPState era, ChainProperty era, HasTrace (CHAIN era) (GenEnv era)) => Property

STAKE INCR

incrementalStakeProp :: forall era. (EraGen era, HasTrace (CHAIN era) (GenEnv era), Default (State (EraRule "PPUP" era))) => Proxy era -> Property

aggregateUtxoCoinByCredential :: forall era. Era era => Map Ptr (Credential 'Staking (Crypto era)) -> UTxO era -> Map (Credential 'Staking (Crypto era)) Coin -> Map (Credential 'Staking (Crypto era)) Coin

Sum up all the Coin for each staking Credential. This function has an incremental analog. See incrementalAggregateUtxoCoinByCredential

splitTrace :: (State s -> State s -> Bool) -> Trace s -> [Trace s]

Split a Trace into several shorter traces. The leftmost Trace (at the front of the list) might be ill-formed, depending on what p does.

forEachEpochTrace :: forall era prop. (EraGen era, Testable prop, HasTrace (CHAIN era) (GenEnv era), Default (State (EraRule "PPUP" era))) => Int -> Word64 -> (Trace (CHAIN era) -> prop) -> Property

Test a property on the first subtracecount sub-Traces that end on an EpochBoundary