Test.Cardano.Ledger.Shelley.Utils

Documentation

mkSeedFromWords :: RawSeed -> Seed

Construct a seed from a bunch of Word64s

We multiply these words by some extra stuff to make sure they contain enough bits for our seed.

mkCertifiedVRF :: (Signable v a, VRFAlgorithm v, ContextVRF v ~ (), Coercible b (CertifiedVRF v a)) => a -> SignKeyVRF v -> b

For testing purposes, create a VRF value

epochFromSlotNo :: SlotNo -> EpochNo

evolveKESUntil

Arguments

:: (KESAlgorithm v, ContextKES v ~ ())

=> SignKeyKES v

-> KESPeriod
Current KES period

-> KESPeriod
Target KES period

-> Maybe (SignKeyKES v)

Try to evolve KES key until specific KES period is reached, given the current KES period.

slotFromEpoch :: EpochNo -> SlotNo

epochSize :: EpochNo -> EpochSize

mkHash :: forall a h. HashAlgorithm h => Int -> Hash h a

mkKeyPair :: forall crypto kd. DSIGNAlgorithm (DSIGN crypto) => RawSeed -> (SignKeyDSIGN (DSIGN crypto), VKey kd crypto)

For testing purposes, generate a deterministic key pair given a seed.

mkKeyPair' :: DSIGNAlgorithm (DSIGN crypto) => RawSeed -> KeyPair kd crypto

For testing purposes, generate a deterministic key pair given a seed.

mkGenKey :: DSIGNAlgorithm (DSIGN crypto) => RawSeed -> (SignKeyDSIGN (DSIGN crypto), VKey kd crypto)

For testing purposes, generate a deterministic genesis key pair given a seed.

mkKESKeyPair :: KESAlgorithm v => RawSeed -> (SignKeyKES v, VerKeyKES v)

For testing purposes, generate a deterministic KES key pair given a seed.

mkVRFKeyPair :: VRFAlgorithm v => RawSeed -> (SignKeyVRF v, VerKeyVRF v)

For testing purposes, generate a deterministic VRF key pair given a seed.

mkAddr :: Crypto crypto => (KeyPair 'Payment crypto, KeyPair 'Staking crypto) -> Addr crypto

runShelleyBase :: ShelleyBase a -> a

testGlobals :: Globals

maxKESIterations :: Word64

unsafeBoundRational :: (HasCallStack, BoundedRational r) => Rational -> r

Convert to a bounded rational type why throwing an error on failure

slotsPerKESIteration :: Word64

testSTS :: forall s. (BaseM s ~ ShelleyBase, STS s, Eq (State s), Show (State s)) => Environment s -> State s -> Signal s -> Either [PredicateFailure s] (State s) -> Assertion

maxLLSupply :: Coin

applySTSTest :: forall s m (rtype :: RuleType). (STS s, RuleTypeRep rtype, m ~ BaseM s) => RuleContext rtype s -> m (Either [PredicateFailure s] (State s))

Apply STS checking assertions.

type GenesisKeyPair crypto = KeyPair 'Genesis crypto

getBlockNonce :: forall era. Era era => Block (BHeader (Crypto era)) era -> Nonce

type ShelleyTest era = (UsesTxBody era, UsesValue era, UsesTxOut era, UsesScript era, UsesAuxiliary era, UsesPParams era, Tx era ~ Tx era, TxSeq era ~ TxSeq era, TxOut era ~ TxOut era, PParams era ~ PParams era, PParamsDelta era ~ PParamsUpdate era, Witnesses era ~ WitnessSet era, Split (Value era), Default (State (EraRule "PPUP" era)), Default (StashedAVVMAddresses era), AnnotatedData (Witnesses era))

type ChainProperty era = (UsesTxOut era, UsesPParams era, UsesValue era, UsesTxBody era, UsesAuxiliary era, Mock (Crypto era), ApplyBlock era, GetLedgerView era, Show (Tx era), Eq (Tx era))

class Split v where

Methods

vsplit :: v -> Integer -> ([v], Coin)

data RawSeed

Constructors

RawSeed !Word64 !Word64 !Word64 !Word64 !Word64

Orphan instances

DSIGNAlgorithm (DSIGN crypto) => Arbitrary (KeyPair kd crypto)

Instance details

Methods

arbitrary :: Gen (KeyPair kd crypto)

shrink :: KeyPair kd crypto -> [KeyPair kd crypto]